% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adjustAccordingToFractions}
\alias{adjustAccordingToFractions}
\title{adjustAccordingToFractions}
\usage{
adjustAccordingToFractions(fractions, singlets)
}
\arguments{
\item{fractions}{Numeric; a numeric vector with length equal to
ncol(singlets) indicating the fractions that each column should be
multiplied with.}

\item{singlets}{Matrix; a counts matrix with cells/samples as columns and
genes as rows.}
}
\description{
This function takes a counts matrix and subsets the columns according to the
indices provided by the idxToSubset argument.
}
\author{
Jason T. Serviss
}
