% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adjustAccordingToFractionsEigen}
\alias{adjustAccordingToFractionsEigen}
\title{adjustAccordingToFractionsEigen}
\usage{
adjustAccordingToFractionsEigen(fractions, subMat)
}
\arguments{
\item{fractions}{Numeric; a numeric vector with length equal to 
\code{ncol(singlets)} indicating the fractions that each column should be 
multiplied with.}

\item{subMat}{Matrix; a counts matrix with cells/samples as columns and 
genes as rows. This matrix should have been previously subset with the 
\code{\link{sampleSingletsEigen}} function so that 
only one singlet per class/cell type is present.}
}
\description{
This function takes a counts matrix and subsets the columns according to the
indices provided by the idxToSubset argument.
}
\author{
Jason T. Serviss
}
