% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\name{averageGroupExpression}
\alias{averageGroupExpression}
\alias{averageGroupExpression}
\title{averageGroupExpression}
\usage{
averageGroupExpression(data, classes, weighted, uncertainty)
}
\arguments{
\item{data}{Singlet expression matrix.}

\item{classes}{A character vector indicating the class of each singlet.}

\item{weighted}{Logical indicating if the group means shoule be weighted with
the uncertainty.}

\item{uncertainty}{A numeric vector indicating the uncertainty if weighted is
TRUE.}
}
\value{
A matrix containing the mean value for each gene for each
classification group.
}
\description{
This output from this function is utilized to represent each group classified
group during swarm optimization.
}
\details{
Calculate the average expression of each gene within each classification
group. Note that typically only singlets are input to this method.
}
\examples{

\dontrun{
singletIdx <- grepl("^s.*", colnames(testCounts))
sngMatrix <- testCounts[ , singletIdx]
sngErcc <- testErcc[ , singletIdx]
cObj <- spCounts(sngMatrix, sngErcc)
select <- spTopMax(cObj, 10)
my.dist <- pearsonsDist(cObj, select)
tsne <- runTsne(my.dist, max_iter=2000)
class <- runMclust(tsne, 50, 11)
averageExp <- averageGroupExpression(cObj, class)
}

}
\author{
Jason T. Serviss
}
\keyword{averageGroupExpression}
