% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMseqSwarm.R
\name{calcResiduals}
\alias{calcResiduals}
\title{calcResiduals}
\usage{
calcResiduals(multiplets, swarm, ...)
}
\arguments{
\item{multiplets}{A CIMseqMultiplets object.}

\item{swarm}{A CIMseqSwarm object.}

\item{...}{additional arguments to pass on}
}
\value{
Residuals (add more description).
}
\description{
Subtitle
}
\details{
Calculates the residuals for each gene and multiplet after deconvolution
based on the spSwarm results.
}
\author{
Jason T. Serviss
}
