% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\name{calculateCost}
\alias{calculateCost}
\alias{calculateCost}
\title{Calculate cost}
\usage{
calculateCost(spCountsMul, spUnsupervised, fractions, distFun, e = NULL,
  cellNumbers = NULL, ...)
}
\arguments{
\item{spCountsMul}{An spCounts object containing multiplets.}

\item{spUnsupervised}{An spUnsupervised object.}

\item{fractions}{The fractions to calculate the cost with.}

\item{distFun}{The distance function to use for the cost calculation.}

\item{e}{Argument to dtsnCellNum distance function.}

\item{cellNumbers}{Argument to dtsnCellNum distance function.}

\item{...}{additional arguments to pass on}
}
\value{
The cost.
}
\description{
Calculates the cost for a multiplet with a input of specific fractions.
}
\details{
Description
}
\examples{

#

}
\author{
Jason T. Serviss
}
\keyword{calculateCost}
