% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateCostArma}
\alias{calculateCostArma}
\title{calculateCostArma}
\usage{
calculateCostArma(oneMultiplet, syntheticMultiplets)
}
\arguments{
\item{oneMultiplet}{Numeric; a numeric vector of counts per million for one
multiplet.}

\item{syntheticMultiplets}{Matrix; a numeric matrix of synthetic multiplets
with samples as columns and genes as rows.}
}
\description{
This function takes a vector of gene counts per million for one multiplet 
that is being deconvoluted and a matrix of synthetic multiplets and 
calculates the cost which is returned to the optimization algorithm during
deconvolution.
}
\author{
Jason T. Serviss
}
