% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMseqSwarm.R
\name{calculateEdgeStats}
\alias{calculateEdgeStats}
\title{calculateEdgeStats}
\usage{
calculateEdgeStats(swarm, singlets, multiplets, theoretical.max = Inf,
  ...)
}
\arguments{
\item{swarm}{A CIMseqSwarm object.}

\item{singlets}{A CIMseqSinglets object.}

\item{multiplets}{A CIMseqMultiplets object.}

\item{theoretical.max}{integer; See \code{\link{estimateCells}}.}

\item{...}{additional arguments to pass on}
}
\value{
CIMseqSwarm connection weights and p-values.
}
\description{
Subtitle
}
\details{
Description
}
\examples{

output <- calculateEdgeStats(
CIMseqSwarm_test, CIMseqSinglets_test, CIMseqMultiplets_test
)

}
\author{
Jason T. Serviss
}
