% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateLogRowMeans}
\alias{calculateLogRowMeans}
\title{calculateLogRowMeans}
\usage{
calculateLogRowMeans(densities)
}
\arguments{
\item{densities}{Numeric; a numeric vector of densities. Typically 
calculated with the \code{\link{calculateCostDensity}}.function.}
}
\description{
This function takes a matrix of poisson densities and calculates the row 
means and subsequently their log10 values.
}
\author{
Jason T. Serviss
}
