% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateLogRowMeansArma}
\alias{calculateLogRowMeansArma}
\title{calculateLogRowMeansArma}
\usage{
calculateLogRowMeansArma(densities)
}
\arguments{
\item{densities}{Numeric; a numeric vector of densities.}
}
\description{
This function takes a matrix of poisson densities and calculates the row 
means and subsequently their log10 values.
}
\author{
Jason T. Serviss
}
