% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{cellNumberPerClass}
\alias{cellNumberPerClass}
\title{cellNumberPerClass}
\usage{
cellNumberPerClass(singlets, multiplets)
}
\arguments{
\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}
}
\value{
A matrix containing the median and mean cell numbers (estimated via
the ERCC fractions) for each classification group.
}
\description{
Calculates median and mean cell estimated cell number per class.
}
\examples{

out <- erccPerClass(CIMseqSinglets_test, CIMseqMultiplets_test)

}
\author{
Jason T. Serviss
}
