% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\docType{methods}
\name{changeCutoff}
\alias{changeCutoff}
\alias{changeCutoff,spSwarm-method}
\title{changeCutoff}
\usage{
changeCutoff(spSwarm, ...)

\S4method{changeCutoff}{spSwarm}(spSwarm, ...)
}
\arguments{
\item{spSwarm}{An spSwarm object.}

\item{...}{additional arguments to pass on}
}
\value{
spSwarm output.
}
\description{
Subtitle
}
\details{
Description
}
\examples{

#use demo data


#run function


}
\author{
Jason T. Serviss
}
\keyword{changeCutoff}

