% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-classes.R, R/CIMseqConnections.R
\docType{class}
\name{CIMseqConnections-class}
\alias{CIMseqConnections-class}
\alias{getData,CIMseqConnections-method}
\alias{CIMseqConnections}
\alias{CIMseqConnections,CIMseqSinglets,CIMseqMultiplets,CIMseqSwarm-method}
\title{CIMseqConnections}
\usage{
\S4method{getData}{CIMseqConnections}(x, n = NULL)

CIMseqConnections(singlets, multiplets, swarm, ...)

\S4method{CIMseqConnections}{CIMseqSinglets,CIMseqMultiplets,CIMseqSwarm}(
  singlets,
  multiplets,
  swarm,
  alpha = 0.01,
  weightCut = 10,
  expectedWeightCut = 0,
  maxCellsPerMultiplet = 4,
  depleted = F,
  multiplet.factor = NA
)
}
\arguments{
\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}

\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{alpha}{numeric; Vector of length 1 specifying the which p-values are 
considered significant.}

\item{weightCut}{integer; Vector of length 1 specifying weights below which 
the p-value should not be calculated (set to non-zero for enrichment).}

\item{expectedWeightCut}{integer; Vector of length 1 specifying expected weights below which 
the p-value should not be calculated (set to non-zero for depletion).}

\item{maxCellsPerMultiplet}{integer; Vector of length 1 specifying the maximum size of a multiplet}

\item{depleted}{logical; Set to TRUE to calculate depletion of connections rather than enrichment}

\item{multiplet.factor}{numeric; Expected average size of multiplets. Used when ERCC controls are not provided.
Should be estimated based on microscopy data. If it is not given, and no ERCC controls are provided, it will be estimated from the data (not recommended).}
}
\value{
CIMseqConnections output.
}
\description{
Subtitle
}
\details{
Description
}
\examples{

#use demo data

}
\author{
Martin Enge
}
