% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{CIMseqMultiplets_test}
\alias{CIMseqMultiplets_test}
\title{CIMseqMultiplets object corresponding to test dataset.}
\format{CIMseqMultiplets
\describe{
    \item{counts}{See \code{\link{CIMseqMultiplets}}.}
    \item{counts.log}{See \code{\link{CIMseqMultiplets}}.}
    \item{counts.cpm}{See \code{\link{CIMseqMultiplets}}.}
    \item{counts.ercc}{See \code{\link{CIMseqMultiplets}}.}
    \item{features}{See \code{\link{CIMseqMultiplets}}.}
}}
\value{
An CIMseqMultiplets object.
}
\description{
CIMseqMultiplets_test a CIMseqMultiplets object.
}
\examples{
data(CIMseqMultiplets_test)

}
