% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-classes.R, R/CIMseqData.R,
%   R/initialize-methods.R, R/show-methods.R
\docType{class}
\name{CIMseqSinglets-class}
\alias{CIMseqSinglets-class}
\alias{getData}
\alias{getData,CIMseqSinglets-method}
\alias{getData<-}
\alias{getData<-,CIMseqSinglets-method}
\alias{CIMseqSinglets}
\alias{CIMseqSinglets,missing-method}
\alias{CIMseqSinglets,matrix-method}
\alias{initialize,CIMseqSinglets-method}
\alias{show,CIMseqSinglets-method}
\title{CIMseqSinglets}
\usage{
getData(x, ...)

\S4method{getData}{CIMseqSinglets}(x, n = NULL)

getData(x, n) <- value

\S4method{getData}{CIMseqSinglets}(x, n = NULL) <- value

CIMseqSinglets(counts, ...)

\S4method{CIMseqSinglets}{missing}(counts, ...)

\S4method{CIMseqSinglets}{matrix}(counts, counts.ercc, dim.red,
  classification, ...)

\S4method{initialize}{CIMseqSinglets}(.Object, ..., counts, counts.log,
  counts.cpm, counts.ercc, dim.red, classification)

\S4method{show}{CIMseqSinglets}(object)
}
\arguments{
\item{x}{CIMseqSinglets; CIMseqSinglets object.}

\item{...}{additional arguments to pass on}

\item{n}{Data to extract from CIMseqSinglets object.}

\item{value}{Replacment value.}

\item{counts}{matrix; Counts matrix with samples as columns and genes as rows.}

\item{counts.ercc}{matrix; A matrix containing ercc spike-in reads.}

\item{dim.red}{matrix; Dimensionality reduced representation of the data.}

\item{classification}{character; Sample classes.}

\item{.Object}{Internal object.}

\item{counts.log}{matrix; Log2 normalized counts per million.}

\item{counts.cpm}{matrix; Normalized counts per million.}

\item{object}{CIMseqSinglets; A CIMseqSinglets to show.}
}
\value{
CIMseqSinglets object.
}
\description{
Subtitle
}
\details{
Imports count, count.ercc, dimensionality reduction, and classification data
to a CIMseqSinglets object for sequenced singlets.
}
\examples{

#setup test data
s <- grepl("^s", colnames(testCounts))
ercc <- grepl("^ERCC\\\\-[0-9]*$", rownames(testCounts))
singlets <- testCounts[!ercc, s]
singletsERCC <- testCounts[ercc, s]

#dimensionality reduction
select <- selectTopMax(singlets, 100)
pdist <- pearsonsDist(singlets, select)
dim.red <- runTsne(pdist)

#classes
classes <- testMeta$cellTypes[match(colnames(singlets), testMeta$sample)]
#setup CIMseqSinglets object
cObj <- CIMseqSinglets(singlets, singletsERCC, dim.red, classes)

}
\author{
Jason T. Serviss
}
