% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{CIMseqSinglets_test}
\alias{CIMseqSinglets_test}
\title{CIMseqSinglets object corresponding to test dataset.}
\format{CIMseqSinglets
\describe{
    \item{counts}{See \code{\link{CIMseqSinglets}}.}
    \item{counts.log}{See \code{\link{CIMseqSinglets}}.}
    \item{counts.cpm}{See \code{\link{CIMseqSinglets}}.}
    \item{counts.ercc}{See \code{\link{CIMseqSinglets}}.}
    \item{dim.red}{See \code{\link{CIMseqSinglets}}.}
    \item{classification}{See \code{\link{CIMseqSinglets}}.}
}}
\value{
An CIMseqSinglets object.
}
\description{
CIMseqSinglets_test a CIMseqSinglets object.
}
\examples{
data(CIMseqSinglets_test)
}
