% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-classes.R, R/CIMseqSwarm.R,
%   R/initialize-methods.R, R/show-methods.R
\docType{class}
\name{CIMseqSwarm-class}
\alias{CIMseqSwarm-class}
\alias{getData,CIMseqSwarm-method}
\alias{CIMseqSwarm}
\alias{CIMseqSwarm,CIMseqSinglets,CIMseqMultiplets-method}
\alias{initialize,CIMseqSwarm-method}
\alias{show,CIMseqSwarm-method}
\title{CIMseqSwarm}
\usage{
\S4method{getData}{CIMseqSwarm}(object, n = NULL)

CIMseqSwarm(singlets, multiplets, ...)

\S4method{CIMseqSwarm}{CIMseqSinglets,CIMseqMultiplets}(singlets,
  multiplets, maxiter = 10, swarmsize = 150,
  nSyntheticMultiplets = 200, seed = 11, norm = TRUE,
  report = FALSE, reportRate = NA, vectorize = FALSE,
  permute = FALSE, ...)

\S4method{initialize}{CIMseqSwarm}(.Object, ..., fractions, costs,
  convergence, stats, singletIdx, arguments)

\S4method{show}{CIMseqSwarm}(object)
}
\arguments{
\item{object}{A CIMseqSwarm object.}

\item{n}{character; Data to extract from CIMseqSwarm object.}

\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}

\item{...}{additional arguments to pass on.}

\item{maxiter}{integer; The maximum optimization iterations.}

\item{swarmsize}{integer; The number of swarm particals.}

\item{nSyntheticMultiplets}{Numeric; Value indicating the number of synthetic
multiplets to generate during deconvolution.}

\item{seed}{integer; Seed to set before running.}

\item{norm}{logical; Indicates if the sum of fractions should equal 1.}

\item{report}{logical; Indicates if additional reporting from the
optimization should be included.}

\item{reportRate}{integer; If report is TRUE, the iteration interval for 
which a report should be generated.}

\item{vectorize}{logical, Argument to \link[pso]{psoptim}.}

\item{permute}{logical; indicates if genes should be permuted before
deconvolution. For use with permutation testing.}

\item{.Object}{Internal object.}

\item{fractions}{matrix; The deconvolution results.}

\item{costs}{numeric; The costs after optimization.}

\item{convergence}{character; The convergence output from \link[pso]{psoptim}.
One value per multiplet.}

\item{stats}{tbl_df; The stats output from \link[pso]{psoptim}}

\item{singletIdx}{list; Indexes indicating singlets that were subset to 
synthesize synthetic multiplets. Facilitates recreation of the synthetic 
multiplets downstream.}

\item{arguments}{list; Arguments passed to the CIMseqSwarm function.}
}
\value{
CIMseqSwarm output.
}
\description{
Subtitle
}
\details{
Description
}
\examples{

#use demo data

}
\author{
Jason T. Serviss
}
