% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{CIMseqSwarm_test}
\alias{CIMseqSwarm_test}
\title{CIMseqSwarm object corresponding to test dataset.}
\format{
CIMseqSwarm
\describe{
 \item{fractions}{See \code{\link{CIMseqSwarm}}.}
 \item{costs}{See \code{\link{CIMseqSwarm}}.}
 \item{convergence}{See \code{\link{CIMseqSwarm}}.}
 \item{singletIdx}{See \code{\link{CIMseqSwarm}}.}
 \item{arguments}{See \code{\link{CIMseqSwarm}}.}
}
}
\value{
CIMseqSwarm object.
}
\description{
CIMseqSwarm_test a CIMseqSwarm object.
}
\examples{
data(CIMseqSwarm_test)

}
