% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\name{classify_louvain}
\alias{classify_louvain}
\title{classify_louvain}
\usage{
classify_louvain(g, weights.col)
}
\arguments{
\item{g}{tbl_graph; A tidygraph graph.}

\item{weights.col}{Bare word; indicating the column name of the column
including the weights.}
}
\description{
Runs the louvain community detection algorithm on the input graph.
}
\author{
Jason T. Serviss
}
