% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotExtras.R
\name{coloursFromTargets}
\alias{coloursFromTargets}
\title{coloursFromTargets}
\usage{
coloursFromTargets(pal, counts, markers, ...)
}
\arguments{
\item{pal}{character; A colour palette with length = length(markers).}

\item{counts}{matrix; A matrix containing counts.}

\item{markers}{character; The markers to evaluate. Must be present in
rownames(counts).}

\item{...}{additional arguments to pass on.}
}
\description{
Diverging color palette with 40 colors.
}
\author{
Jason T. Serviss
}
\keyword{coloursFromTargets}
