% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{constructGraph}
\alias{constructGraph}
\title{constructGraph}
\usage{
constructGraph(data, distCut)
}
\arguments{
\item{data}{Tibble; Expected columns are "from" and "to", indicating the
edges and "dist" indicating the distance between the samples.}

\item{distCut}{numeric; A Euclidean distance at which NN are removed.}
}
\description{
Constructs a graph from a data frame or tibble.
}
\author{
Jason T. Serviss
}
