% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExtras.R
\name{convertToERCC}
\alias{convertToERCC}
\title{convertToERCC}
\usage{
convertToERCC(ercc, singlets, multiplets)
}
\arguments{
\item{ercc}{The left axis values. Passes as ".".}

\item{singlets}{CIMseqSinglets; An CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; An CIMseqMultiplets object.}
}
\description{
A function to facilitate calculation of the second axis of the plotCounts
type "ercc" plot.
}
\author{
Jason T. Serviss
}
\keyword{convertToERCC}
