% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{costFor}
\alias{costFor}
\title{costFor}
\usage{
costFor(oneMultiplet, singletSubset, fractions, n)
}
\arguments{
\item{oneMultiplet}{Integer; a integer vector of rounded counts per million
for one multiplet.}

\item{singletSubset}{Matrix; Numeric matrix with the preallocated singlets.
Each of the n synthetic multiplets should be stacked, i.e. rbind.}

\item{fractions}{Numeric; a numeric vector with length equal to
ncol(singlets) indicating the fractions that each column should be
multiplied with.}

\item{n}{Integer; length 1 integer indicating the number of synthetic
multiplets to generate.}
}
\description{
Wrapper for deconvolution cost function using the Armadillo C++ library.
}
\author{
Jason T. Serviss
}
