% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{costNegSumArma}
\alias{costNegSumArma}
\title{costNegSumArma}
\usage{
costNegSumArma(means)
}
\arguments{
\item{means}{Numeric; a numeric vector of log10 row means.}
}
\description{
This function takes a numeric vector and calculates the negative sum.
}
\author{
Jason T. Serviss
}
