% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\name{countsPerClass}
\alias{countsPerClass}
\title{countsPerClass}
\usage{
countsPerClass(spCountsSng, spUnsupervised)
}
\arguments{
\item{spCountsSng}{An spCounts object containing singlets.}

\item{spUnsupervised}{An spUnsupervised object.}
}
\value{
A matrix containing the median value for each gene for each
   classification group.
}
\description{
Calculates median counts per class.
}
\examples{

output <- countsPerClass(test_spCountsSng, test_spUnsupervised)

}
\author{
Jason T. Serviss
}
\keyword{countsPerClass}
