% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpmEigen}
\alias{cpmEigen}
\title{cpmEigen}
\usage{
cpmEigen(counts)
}
\arguments{
\item{counts}{Matrix; a counts matrix. Must be formated as numeric.}
}
\description{
Takes a numeric matrix of counts and calculates counts per million.
}
\author{
Jason T. Serviss
}
