% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyFunctions.R
\name{divide_by}
\alias{divide_by}
\title{divide_by}
\usage{
divide_by()
}
\description{
Facilitates division in pipes and "avoids wrong number of arguments to"
complaint by check. Adopted from magrittr:
\link{https://github.com/tidyverse/magrittr/blob/master/R/aliases.R#L93-L96}
}
\author{
Jason T. Serviss
}
