% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExtras.R
\name{draw_legend}
\alias{draw_legend}
\title{draw_legend}
\usage{
draw_legend(l)
}
\arguments{
\item{l}{A grob containing a legend.}
}
\description{
Helper function for plotSwarmCircos.
}
\author{
Jason T. Serviss
}
