% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\name{dtsnCellNum}
\alias{dtsnCellNum}
\title{dtsnCellNum}
\usage{
dtsnCellNum(fractions, cellTypes, oneMultiplet, e, cellNumber, ...)
}
\arguments{
\item{fractions}{numeric; Fractions to adjust.}

\item{cellTypes}{matrix; Mean cell type gene expression.}

\item{oneMultiplet}{numeric; Gene expression values for multiplet.}

\item{e}{numeric; epsilon}

\item{cellNumber}{matrix? Estimated number of cells per multiplet.}

\item{...}{Additional arguments to pass.}
}
\description{
Cost function for optimization.
}
\author{
Jason T. Serviss
}
