% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\name{estimateTotalConnections}
\alias{estimateTotalConnections}
\title{estimateTotalConnections}
\usage{
estimateTotalConnections(spCountsSng, spCountsMul)
}
\arguments{
\item{spCountsSng}{An spCounts object containing singlets.}

\item{spCountsMul}{An spCounts object containing multiplets.}
}
\value{
A table with the expected frequencies..
}
\description{
Estimates the total number of connections in the multiplets.
}
\examples{

expectedInteractionFreq(testUns)
}
\author{
Jason T. Serviss
}
\keyword{estimateTotalConnections}
