% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expCounts.R
\docType{data}
\name{expCounts}
\alias{expCounts}
\title{Fetal pancreas counts data.}
\format{matrix
\describe{
    \item{rownames}{Gene names}
    \item{colnames}{Samples. Singlets prefixed with "s" and multiplets "m".}
}}
\usage{
expCounts
}
\value{
Matrix of counts.
}
\description{
Fetal pancreas counts data.
}
\examples{
data(expCounts)

}
