% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{expData}
\alias{expData}
\title{Experimental spatial scRNAseq data.}
\format{spCounts
#\describe{
#\item{rownames}{Groups}
#\item{colnames}{dimension number}
#}}
\usage{
expData
}
\value{
Dataset of class spCounts
}
\description{
Experimental data.
}
\examples{
expData

}
