% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\name{expectedInteractionFreq}
\alias{expectedInteractionFreq}
\alias{expectedInteractionFreq}
\title{expectedInteractionFreq}
\usage{
expectedInteractionFreq(spUnsupervised)
}
\arguments{
\item{spUnsupervised}{An spUnsupervised object.}
}
\value{
A table with the expected frequencies..
}
\description{
Calculates the expected interaction frequency between cell types.
}
\examples{

expectedInteractionFreq(testUns)
}
\author{
Jason T. Serviss
}
\keyword{expectedInteractionFreq}
