% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expErcc.R
\docType{data}
\name{expErcc}
\alias{expErcc}
\title{Fetal pancreas ERCC counts data.}
\format{matrix
\describe{
    \item{rownames}{ERCC spike-in names.}
    \item{colnames}{Samples. Singlets prefixed with "s" and multiplets "m".}
}}
\usage{
expErcc
}
\value{
Matrix of ERCC counts.
}
\description{
Fetal pancreas ERCC counts data.
}
\examples{
data(expErcc)

}
