% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixNegInfArma}
\alias{fixNegInfArma}
\title{fixNegInfArma}
\usage{
fixNegInfArma(means)
}
\arguments{
\item{means}{Numeric; a numeric vector of log10 row means.}
}
\description{
This function takes a numeric vector and replaces -Inf values with 
-323.0052. Note: since log10(10^-324) gives -Inf but log10(10^-323) 
gives -323.0052
}
\author{
Jason T. Serviss
}
