% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExtras.R
\name{g_legend}
\alias{g_legend}
\title{g_legend}
\usage{
g_legend(a.gplot)
}
\arguments{
\item{a.gplot}{A ggplot.}
}
\description{
Helper function for plotSwarmCircos.
}
\author{
Jason T. Serviss
}
\keyword{g_legend}
