% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateSyntheticMultipletsEigen}
\alias{generateSyntheticMultipletsEigen}
\title{generateSyntheticMultipletsEigen}
\usage{
generateSyntheticMultipletsEigen(singlets, classes, fractions, n)
}
\arguments{
\item{singlets}{Matrix; a counts matrix with cells/samples as columns and 
genes as rows.}

\item{classes}{Character; a character vector of classes with length equal to
the number of cells for which counts exist.}

\item{fractions}{Numeric; a numeric vector with length equal to 
\code{ncol(singlets)} indicating the fractions that each column should be 
multiplied with.}

\item{n}{Integer; length 1 integer indicating the number of synthetic 
multiplets to generate.}
}
\description{
Wrapper for deconvolution cost function using the Eigen C++ library.
}
\author{
Jason T. Serviss
}
