% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMseqSwarm.R
\name{getCellsForMultiplet}
\alias{getCellsForMultiplet}
\alias{getCellsForMultiplet,CIMseqSwarm-method}
\title{getCellsForMultiplet}
\usage{
getCellsForMultiplet(swarm, ...)

\S4method{getCellsForMultiplet}{CIMseqSwarm}(swarm, singlets, multiplets, multipletName = NULL, ...)
}
\arguments{
\item{swarm}{A CIMseqSwarm object.}

\item{...}{additional arguments to pass on}

\item{singlets}{A CIMseqSinglets object.}

\item{multiplets}{A CIMseqMultiplets object.}

\item{multipletName}{character; The name of the multiplet of interest.}
}
\value{
Edge names.
}
\description{
Returns the names of the cells detected in a multiplet.
}
\details{
Description
}
\examples{

output <- getCellsForMultiplet(
CIMseqSwarm_test, CIMseqSinglets_test, CIMseqMultiplets_test,
"m.NJB00204.G04"
)

}
\author{
Jason T. Serviss
}
