% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMseqSwarm.R
\docType{methods}
\name{getEdgesForMultiplet}
\alias{getEdgesForMultiplet}
\alias{getEdgesForMultiplet,CIMseqSwarm-method}
\title{getEdgesForMultiplet}
\usage{
getEdgesForMultiplet(swarm, ...)

\S4method{getEdgesForMultiplet}{CIMseqSwarm}(swarm, edge.cutoff, multiplet,
  ...)
}
\arguments{
\item{swarm}{A CIMseqSwarm object.}

\item{...}{additional arguments to pass on}

\item{edge.cutoff}{numeric; The minimum fraction to consider (?).}

\item{multiplet}{character; The name of the multiplet of interest.}
}
\value{
Edge names.
}
\description{
Returns the names of the edges are associated with a multiplet.
}
\details{
Description
}
\examples{

output <- getEdgesForMultiplet(CIMseqSwarm_test, 1/10.5, "m.NJB00204.G04")

}
\author{
Jason T. Serviss
}
