% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{getKNN}
\alias{getKNN}
\title{getKNN}
\usage{
getKNN(pca, pcs, k, mask)
}
\arguments{
\item{pca}{Matrix; Principal components as columns, samples as rows.}

\item{pcs}{Integer; Length 1 vector indicating the max principal component to
retain.}

\item{k}{Integer; Number of nearest neighbors.}

\item{mask}{Integer, Principal components to mask.}
}
\description{
Calculates Euclidean distance in PCA space.
}
\author{
Jason T. Serviss
}
