% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\docType{methods}
\name{getMultipletsForEdge}
\alias{getMultipletsForEdge}
\alias{getMultipletsForEdge}
\alias{getMultipletsForEdge,spSwarm-method}
\title{getMultipletsForEdge}
\usage{
getMultipletsForEdge(spSwarm, ...)

\S4method{getMultipletsForEdge}{spSwarm}(spSwarm, edge.cutoff, clust1, clust2,
  ...)
}
\arguments{
\item{spSwarm}{An spSwarm object.}

\item{...}{additional arguments to pass on}

\item{edge.cutoff}{The minimum fraction to consider (?).}

\item{clust1}{A character vector of length 1 indicating one node which forms
an edge with clust2 for which multiplets contributing to that edge will be
reported.}

\item{clust2}{A character vector of length 1 indicating one node which forms
an edge with clust1 for which multiplets contributing to that edge will be
reported.}

\item{object}{spRSwarm object.}

\item{.Object}{Internal object.}
}
\value{
Multiplet names.
}
\description{
Returns the names of the multiplets that are associated with an edge.
}
\details{
Description
}
\examples{

#use demo data


#run function


}
\author{
Jason T. Serviss
}
\keyword{getMultipletsForEdge}
