% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\name{jaccardSimilarity}
\alias{jaccardSimilarity}
\title{jaccardSimilarity}
\usage{
jaccardSimilarity(g, prune = 1/15, seed = 7823)
}
\arguments{
\item{g}{tbl_graph; A tidygraph graph.}

\item{seed}{integer; a seed to set before running the community detection.}
}
\description{
Calculates the Jaccard index pairwise for graph nodes..
}
\author{
Jason T. Serviss
}
