% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyFunctions.R
\name{matrix_to_tibble}
\alias{matrix_to_tibble}
\title{matrix_to_tibble}
\usage{
matrix_to_tibble(data, rowname = "rowname", drop = FALSE)
}
\arguments{
\item{data}{matrix; The matrix to be converted.}

\item{rowname}{character; Length 1 vector indicating the colname that
rownames should have upon tibble conversion.}

\item{drop}{logical; indicated if rownames should be dropped.
Default = FALSE.}
}
\description{
Converts a matrix to a tibble without removing rownames.
}
\examples{

m <- matrix(rnorm(20), ncol = 2, dimnames = list(letters[1:10], LETTERS[1:2]))
output <- matrix_to_tibble(m)

}
\author{
Jason T. Serviss
}
\keyword{matrix_to_tibble}
