% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{means.dim.red}
\alias{means.dim.red}
\title{means.dim.red}
\usage{
means.dim.red(data, classes)
}
\arguments{
\item{data}{Singlet 2D tsne.}

\item{classes}{A character vector indicating the class of each singlet.}
}
\value{
A matrix containing the mean value for each gene for each
   classification group.
}
\description{
Calculates the x and y coordinates of the mean of each classified group in 
the dimensionality reduced data.
}
\details{
This method is typically only used in conjunction with plotting. It
calculates the 2 dimensional location of the mean of each classified group
in the supplied unsupervised dimensionality reduction (t-SNE) data
representation.
}
\examples{

means <- means.dim.red(
  getData(CIMseqSinglets_test, "dim.red"),
  getData(CIMseqSinglets_test, "classification")
)

}
\author{
Jason T. Serviss
}
