% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multipletSums}
\alias{multipletSums}
\title{multipletSums}
\usage{
multipletSums(adjusted)
}
\arguments{
\item{adjusted}{Matrix; a counts matrix with cells/samples as columns and
genes as rows.}
}
\description{
This function takes a counts matrix and calculates the row sums. Output is
subsequently rounded for integration downstream.
}
\author{
Jason T. Serviss
}
