% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multipletSumsEigen}
\alias{multipletSumsEigen}
\title{multipletSumsEigen}
\usage{
multipletSumsEigen(singlets)
}
\arguments{
\item{singlets}{matrix; A counts matrix with cells/samples as columns and 
genes as rows. This matrix should have been previously subset with the 
\code{\link{subsetSingletsEigen}} function so that
only one singlet per class/cell type is present. Furthermore, it should have
already been asjusted by the fractions using the 
\code{\link{adjustAccordingToFractionsEigen}} function.}
}
\description{
This function takes a counts matrix and calculates the row sums.
}
\author{
Jason T. Serviss
}
