% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyFunctions.R
\name{multiply_by}
\alias{multiply_by}
\title{multiply_by}
\usage{
multiply_by()
}
\description{
Facilitates multiplication in pipes and "avoids wrong number of arguments to"
complaint by check. Adopted from magrittr:
\link{https://github.com/tidyverse/magrittr/blob/master/R/aliases.R#L82-L85}
}
\author{
Jason T. Serviss
}
