% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{namedListToTibble}
\alias{namedListToTibble}
\title{namedListToTibble}
\usage{
namedListToTibble(l)
}
\arguments{
\item{l}{List. The list to be converted.}
}
\description{
Converts a named list to a long data frame.
}
\examples{

l <- list(a=LETTERS[1:10], b=letters[1:5])
output <- namedListToTibble(l)

}
\author{
Jason T. Serviss
}
\keyword{namedListToTibble}
