% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalizeFractions}
\alias{normalizeFractions}
\title{normalizeFractions}
\usage{
normalizeFractions(fractions)
}
\arguments{
\item{fractions}{Numeric; a numeric vector.}
}
\description{
Takes a numeric vector and scales to [0, 1] by dividing with its sum.
}
\author{
Jason T. Serviss
}
