% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyFunctions.R
\name{normalizeVec}
\alias{normalizeVec}
\title{normalizeVec}
\usage{
normalizeVec(vec)
}
\arguments{
\item{vec}{numeric; A numeric vector.}
}
\description{
Normalizes a vector (x) using x - min(x) / max(x) - min(x).
}
\examples{

normalizeVec(rnorm(100))

}
\author{
Jason T. Serviss
}
\keyword{normalizeVec}
