% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{pcByVarPercent}
\alias{pcByVarPercent}
\title{pcByVarPercent}
\usage{
pcByVarPercent(pca, cutoff)
}
\arguments{
\item{pca}{Output from \code{\link[gmodels]{fast.prcomp}} or
\code{\link[stats]{prcomp}}.}

\item{cutoff}{numeric; Select PCs containing variance greater or equal to
this cutoff.}
}
\value{
The maximum PC that contains variance specified by the cutoff
argument.
}
\description{
Runs a principal component analysis using \code{\link[gmodels]{fast.prcomp}}.
}
\author{
Jason T. Serviss
}
