% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{pearsonsDist}
\alias{pearsonsDist}
\title{pearsonsDist}
\usage{
pearsonsDist(cpm, select)
}
\arguments{
\item{cpm}{matrix; Counts per million.}

\item{select}{A numeric vector indicating the indexes of genes to include.}
}
\value{
A matrix containing the mean value for each gene for each
   classification group.
}
\description{
Calculates the x and y coordinates of the mean of each classified group.
}
\details{
This method is typically only used in conjunction with plotting. It
calculates the 2 dimensional location of the mean of each classified group
in the supplied unsupervised dimensionality reduction (t-SNE) data
representation.
}
\examples{

d <- pearsonsDist(getData(CIMseqSinglets_test, "counts.cpm"), 1:2000)

}
\author{
Jason T. Serviss
}
\keyword{pearsonsDist}
