% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMseqConnections.R
\name{plotConnectionCircos}
\alias{plotConnectionCircos}
\alias{plotConnectionCircos,CIMseqConnections-method}
\title{plotConnectionCircos}
\usage{
plotConnectionCircos(connections, ...)

\S4method{plotConnectionCircos}{CIMseqConnections}(
  connections,
  classOrder = NULL,
  label.cex = 1,
  legend = TRUE,
  pal = colorRampPalette(c("grey95", viridis::viridis(1)))(120)[40:120],
  nonSigCol = "grey95",
  h.ratio = 0.5,
  drawFractions = T,
  ...
)
}
\arguments{
\item{connections}{CIMseqConnections; A CIMseqConnections object.}

\item{...}{additional arguments to pass on.}

\item{classOrder}{character; Order of the cell types / classes on the circos.}

\item{label.cex}{numeric; Vector of length 1 between [0, 1] indicating the
size of the cell type labels.}

\item{legend}{logical; indicates if the legends should be plotted.}

\item{pal}{character; A vector including the colour pallete for the score 
colours.}

\item{nonSigCol}{character; Vector of length 1 indicating the colours for 
non-significant connections.}

\item{h.ratio}{numeric; Vector of length 1 indicating the curvature of links}

\item{drawFraction}{logical; Vector of length 1 indicating the colours for 
non-significant connections.}
}
\value{
A ggplot object.
}
\description{
Subtitle
}
\details{
Description
}
\author{
Martin Enge
}
