% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlot.R
\docType{methods}
\name{plotCounts}
\alias{plotCounts}
\alias{plotCounts}
\alias{plotCounts,spCounts-method}
\title{plotCounts}
\usage{
plotCounts(spCountsSng, ...)

\S4method{plotCounts}{spCounts}(spCountsSng, spCountsMul, type = "ercc",
  markers = NULL, ...)
}
\arguments{
\item{spCountsSng}{spCounts; An spCounts object containing singlets.}

\item{...}{additional arguments to pass on.}

\item{spCountsMul}{spCounts; An spCounts object containing multiplets.}

\item{type}{character; Can be "ercc", "markers". Default is "ercc".}

\item{markers}{character; A vector with the 2 markers to plot.}
}
\value{
The ggplot2 object with the t-SNE results plotted on the x and y
 axis. The plot can be modified by adding geoms, themes, etc. in the normal
 manner with ggplot2. See examples or the plotting vignette for further help.
}
\description{
Plot method for spCounts objects.
}
\examples{

#use demo data
s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])
cObjMul <- spCounts(testCounts[, !s], testErcc[, !s])

#ERCC plot
p <- plotCounts(cObjSng, cObjMul) + geom_jitter()

#Markers plot
markers <- c("a1", "a10")
p <- plotCounts(cObjSng, cObjMul, type = "markers", markers = markers) +
geom_point()
}
\author{
Jason T. Serviss
}
\keyword{plotCounts}
