% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotCounts.R
\docType{methods}
\name{plotCountsERCC}
\alias{plotCountsERCC}
\alias{plotCountsERCC,spCounts,spCounts-method}
\title{plotCountsERCC}
\usage{
plotCountsERCC(spCountsSng, spCountsMul, ...)

\S4method{plotCountsERCC}{spCounts,spCounts}(spCountsSng, spCountsMul, ...)
}
\arguments{
\item{spCountsSng}{spCounts; An spCounts object containing singlets.}

\item{spCountsMul}{spCounts; An spCounts object containing multiplets.}

\item{...}{additional arguments to pass on.}
}
\value{
A ggplot2 object containing the plot. See examples or the plotting
 vignette for further help.
}
\description{
Plot method for spCounts objects to display ERCC fraction and estimated cell
number per sample.
}
\examples{

#use demo data
s <- grepl("^s", colnames(testCounts))
ercc <- grepl("^ERCC\\\\-[0-9]*$", rownames(testCounts))
cObjSng <- spCounts(testCounts[!ercc, s], testCounts[ercc, s])
cObjMul <- spCounts(testCounts[!ercc, !s], testCounts[ercc, !s])

#ERCC plot
p <- plotCountsERCC(cObjSng, cObjMul)

}
\author{
Jason T. Serviss
}
\keyword{plotCountsERCC}
