% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotCounts.R
\docType{methods}
\name{plotCountsMarkers}
\alias{plotCountsMarkers}
\alias{plotCountsMarkers,spCounts,spCounts-method}
\title{plotCountsMarkers}
\usage{
plotCountsMarkers(spCountsSng, spCountsMul, ...)

\S4method{plotCountsMarkers}{spCounts,spCounts}(spCountsSng, spCountsMul,
  markers = NULL, ...)
}
\arguments{
\item{spCountsSng}{spCounts; An spCounts object containing singlets.}

\item{spCountsMul}{spCounts; An spCounts object containing multiplets.}

\item{...}{additional arguments to pass on.}

\item{markers}{character; A vector with the 2 markers to plot.}
}
\value{
A ggplot2 object containing the plot. See examples or the plotting
 vignette for further help.
}
\description{
Plot method for spCounts objects to display "markers", typically genes
thought to be discreetly expressed in one cell type, in all samples.
}
\examples{

#use demo data
s <- grepl("^s", colnames(testCounts))
ercc <- grepl("^ERCC\\\\-[0-9]*$", rownames(testCounts))
cObjSng <- spCounts(testCounts[!ercc, s], testCounts[ercc, s])
cObjMul <- spCounts(testCounts[!ercc, !s], testCounts[ercc, !s])

#Markers plot
markers <- c("ACTB", "GAPDH")
p <- plotCountsMarkers(cObjSng, cObjMul, markers)
}
\author{
Jason T. Serviss
}
\keyword{plotCountsMarkers}
