% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExtras.R
\docType{methods}
\name{plotData}
\alias{plotData}
\alias{plotData,gg-method}
\title{plotData}
\usage{
plotData(plot, ...)

\S4method{plotData}{gg}(plot, ...)
}
\arguments{
\item{plot}{A ggplot object of class "gg" "ggplot".}

\item{...}{additional arguments to pass on.}
}
\value{
A tibble containing the plot data.
}
\description{
Returns the data used to build plots for spCounts, spUnsupervised, and
spSwarm objects.
}
\examples{

p <- plotCountsERCC(CIMseqSinglets_test, CIMseqMultiplets_test)
plotData(p)

}
\author{
Jason T. Serviss
}
