% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{plotErccFrac}
\alias{plotErccFrac}
\alias{plotErccFrac,spCounts-method}
\title{Plot Ercc Fraction}
\usage{
plotErccFrac(x, ...)

\S4method{plotErccFrac}{spCounts}(x, ...)
}
\arguments{
\item{...}{Additional arguments to pass on}

\item{counts}{Counts object.}
}
\value{
Ercc fraction plot.
}
\description{
Subtitle
}
\details{
Imports count and count.ercc data to a sp.scRNAseq object.
}
\examples{

#use demo data
data(Doublet_project_data)
x <- importCounts(counts, counts.ercc)

#run function
plotErccFrac(x)

}
\author{
Jason T. Serviss
}
\keyword{plotErccFrac}

