% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{plotMarkers}
\alias{plotMarkers}
\alias{plotMarkers,spCounts-method}
\title{Plot Markers}
\usage{
plotMarkers(x, ...)

\S4method{plotMarkers}{spCounts}(x, marker1, marker2, ...)
}
\arguments{
\item{...}{Additional arguments to pass on}

\item{marker1}{Marker to plot on the x-axis.}

\item{marker2}{Marker to plot on the y-axis.}

\item{counts}{Counts object.}
}
\value{
Ercc fraction plot.
}
\description{
Subtitle
}
\details{
Imports count and count.ercc data to a sp.scRNAseq object.
}
\examples{

#use demo data
data(Doublet_project_data)
x <- importCounts(counts, counts.ercc)

#run function
plotMarkers(x, marker1="A1BG", marker2="A1CF")

}
\author{
Jason T. Serviss
}
\keyword{plotMarkers}

