% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlot.R
\docType{methods}
\name{plotSwarm}
\alias{plotSwarm}
\alias{plotSwarm}
\alias{plotSwarm,spSwarm-method}
\title{plotSwarm}
\usage{
plotSwarm(x, ...)

\S4method{plotSwarm}{spSwarm}(x, y, z = NULL, type = NULL,
  layout = "tsne", loop = FALSE, markers = NULL, edge.cutoff = 0,
  min.pval = 1, min.num.edges = 0, label.cutoff = 1, ...)
}
\arguments{
\item{x}{An spSwarm object.}

\item{...}{additional arguments to pass on.}

\item{y}{An spUnsupervised object.}

\item{z}{An spCounts object with singlets.}

\item{type}{Can be "ercc", "markers", or .......}

\item{layout}{Specify "tsne" for overlaying connections over the tSNE or
igraph for graph layout.}

\item{loop}{Logical; TRUE if "self" connections should be plotted.}

\item{markers}{Not currently implemented.}

\item{edge.cutoff}{The minimum fraction to consider (?).}

\item{min.pval}{Minimum p-value to report.}

\item{min.num.edges}{Minimum number of observed edges to report a connection.}

\item{label.cutoff}{The number of labels to show in the residuals plot.}
}
\value{
The spPlot function returns an object of class spCounts.
}
\description{
Subtitle
}
\details{
Description.
}
\examples{

#use demo data

#run function

}
\author{
Jason T. Serviss
}
\keyword{plotSwarm}
