% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSwarm.R
\name{plotSwarmBarBase}
\alias{plotSwarmBarBase}
\alias{plotSwarmBarBase,CIMseqSwarm,CIMseqSinglets,CIMseqMultiplets-method}
\title{plotSwarmBarBase}
\usage{
plotSwarmBarBase(swarm, singlets, multiplets, ...)

\S4method{plotSwarmBarBase}{CIMseqSwarm,CIMseqSinglets,CIMseqMultiplets}(swarm, singlets, multiplets, ...)
}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}

\item{...}{additional arguments to pass on.}
}
\value{
A ggplot object with the base data.
}
\description{
Not exported. Provides the base plot for all spSwarm bar type plots.
}
\author{
Jason T. Serviss
}
