% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSwarm.R
\docType{methods}
\name{plotSwarmEdgeBar}
\alias{plotSwarmEdgeBar}
\alias{plotSwarmEdgeBar,CIMseqSwarm,CIMseqSinglets,CIMseqMultiplets-method}
\title{plotSwarmEdgeBar}
\usage{
plotSwarmEdgeBar(swarm, singlets, multiplets, ...)


  \S4method{plotSwarmEdgeBar}{CIMseqSwarm,CIMseqSinglets,CIMseqMultiplets}(swarm,
  singlets, multiplets, ...)
}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}

\item{...}{additional arguments to pass on.}
}
\value{
A ggplot object.
}
\description{
plotSwarmEdgeBar
}
\author{
Jason T. Serviss
}
