% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSwarm.R
\name{plotSwarmGraph}
\alias{plotSwarmGraph}
\alias{plotSwarmGraph,CIMseqSwarm,CIMseqSinglets,CIMseqMultiplets-method}
\title{plotSwarmGraph}
\usage{
plotSwarmGraph(swarm, singlets, multiplets, ...)

\S4method{plotSwarmGraph}{CIMseqSwarm,CIMseqSinglets,CIMseqMultiplets}(swarm, singlets, multiplets, ...)
}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}

\item{...}{additional arguments to pass on.}
}
\value{
The spPlot function returns an object of class spCounts.
}
\description{
Description.
}
\details{
Details.
}
\examples{

p <- plotSwarmGraph(
CIMseqSwarm_test, CIMseqSinglets_test, CIMseqMultiplets_test
)

}
\author{
Jason T. Serviss
}
\keyword{plotSwarmGraph}
