% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSwarm.R
\docType{methods}
\name{plotSwarmPbar}
\alias{plotSwarmPbar}
\alias{plotSwarmPbar,CIMseqSwarm-method}
\title{plotSwarmPbar}
\usage{
plotSwarmPbar(swarm, ...)

\S4method{plotSwarmPbar}{CIMseqSwarm}(swarm, ...)
}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{...}{additional arguments to pass on.}
}
\value{
A ggplot object.
}
\description{
plotSwarmPbar
}
\author{
Jason T. Serviss
}
\keyword{plotSwarmPbar}
