% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlot.R
\docType{methods}
\name{plotUnsupervised}
\alias{plotUnsupervised}
\alias{plotUnsupervised}
\alias{plotUnsupervised,spUnsupervised-method}
\title{plotUnsupervised}
\usage{
plotUnsupervised(spUnsupervised, ...)

\S4method{plotUnsupervised}{spUnsupervised}(spUnsupervised, spCounts,
  type = "clusters", markers = NULL, pal = NULL, ...)
}
\arguments{
\item{spUnsupervised}{spUnsupervised; An spUnsupervised object.}

\item{...}{additional arguments to pass on.}

\item{spCounts}{spCounts; An spCounts object containing singlets.}

\item{type}{Not implemented.}

\item{markers}{character; A vector with markers to be included plot.}

\item{pal}{character; A palette of colors with length(pal) = length(markers).}
}
\value{
The ggplot2 object with the t-SNE results plotted on the x and y
 axis. The plot can be modified by adding geoms, themes, etc. in the normal
manner with ggplot2. See examples or the plotting vignette for further help.
}
\description{
Plot method for spUnsupervised objects.
}
\examples{

#use demo data
s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])
uObj <- testUns

#Clusters plot
p <- plotUnsupervised(uObj, cObjSng) +
geom_point()

#Markers plot
markers <- c("a1", "a10")
pal <- c("red", "blue")
p <- plotUnsupervised(uObj, cObjSng, markers = markers, pal = pal)

}
\author{
Jason T. Serviss
}
\keyword{plotUnsupervised}
