% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts.R
\docType{methods}
\name{plotUnsupervisedClass}
\alias{plotUnsupervisedClass}
\alias{plotUnsupervisedClass,CIMseqSinglets-method}
\title{plotUnsupervisedClass}
\usage{
plotUnsupervisedClass(singlets, ...)

\S4method{plotUnsupervisedClass}{CIMseqSinglets}(singlets, multiplets, ...)
}
\arguments{
\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{...}{additional arguments to pass on.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}
}
\value{
The ggplot2 object containing the plot. See examples or the plotting
 vignette for further help.
}
\description{
Plot method for spUnsupervised objects to highlight dimensionality reduction
and classification results.
}
\examples{

p <- plotUnsupervisedClass(CIMseqSinglets_test, CIMseqMultiplets_test)

}
\author{
Jason T. Serviss
}
