% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotUnsupervised.R
\name{plotUnsupervisedData}
\alias{plotUnsupervisedData}
\title{plotUnsupervisedData}
\usage{
plotUnsupervisedData(spUnsupervised, spCountsSng, markers = NULL,
  pal = NULL, ...)
}
\arguments{
\item{spUnsupervised}{spUnsupervised; An spUnsupervised object.}

\item{spCountsSng}{spCounts; An spCounts object containing singlets.}

\item{markers}{character; A vector with the markers to plot. Must be present
in rownames of counts.}

\item{pal}{character; The colour palette to use.}

\item{...}{additional arguments to pass on.}
}
\value{
A tibble with columns:
}
\description{
Assembles all data for plotUnsupervised plots.
}
\examples{

plotUnsupervisedData(test_spUnsupervised, test_spCountsSng)

}
\author{
Jason T. Serviss
}
\keyword{plotUnsupervisedData}
