% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poissonSampleArma}
\alias{poissonSampleArma}
\title{poissonSampleArma}
\usage{
poissonSampleArma(rsRcpp)
}
\arguments{
\item{rsRcpp}{Integer; vector of rounded row sums.}
}
\description{
This function takes the rowSums calculated by the 
\code{\link{multipletSumsArma}}.function and randomly samples 1 value for 
each input using the Poisson distribution and the input value as lambda.
}
\author{
Jason T. Serviss
}
