% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{preallocCost}
\alias{preallocCost}
\title{preallocCost}
\usage{
preallocCost(oneMultiplet, singletSubset, fractions)
}
\arguments{
\item{oneMultiplet}{Numeric; a numeric vector of counts per million for one
multiplet.}

\item{singletSubset}{Matrix; Numeric matrix with the preallocated singlets. 
Each of the n synthetic multiplets should be stacked, i.e. rbind.}

\item{fractions}{Numeric; a numeric vector with length equal to 
ncol(singlets) indicating the fractions that each column should be 
multiplied with.}
}
\description{
Calculates cost with a preallocated matrix of subsetted singlets.
}
\author{
Jason T. Serviss
}
