% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExtras.R
\name{processMarkers}
\alias{processMarkers}
\title{processMarkers}
\usage{
processMarkers(counts.log, markers)
}
\arguments{
\item{counts.log}{matrix; A matrix containing log2(cpm).}

\item{markers}{character; The markers to process. Must be present in
rownames(counts.log).}
}
\description{
Helper function for plotCountsMarkers and plotUnsupervisedMarkers. Gathers
and returns data in an expected format.
}
\author{
Jason T. Serviss
}
\keyword{processMarkers}
