% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\docType{methods}
\name{pySwarm}
\alias{pySwarm}
\alias{pySwarm,spCounts-method}
\title{pySwarm}
\usage{
pySwarm(spCounts, ...)

\S4method{pySwarm}{spCounts}(spCounts, spUnsupervised, ...)
}
\arguments{
\item{spCounts}{An spCounts object.}

\item{...}{additional arguments to pass on}

\item{spUnsupervised}{An spUnsupervised object.}
}
\value{
PySwarm output.
}
\description{
Subtitle
}
\details{
Description
}
\examples{

#use demo data


#run function


}
\author{
Jason T. Serviss
}
\keyword{pySwarm}

