% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\name{removeOutliers}
\alias{removeOutliers}
\title{removeOutliers}
\usage{
removeOutliers(g, classificationMethod, threshold = 1)
}
\arguments{
\item{g}{tbl_graph; A tidygraph graph.}

\item{classificationMethod}{Bare word; Name of the column containing the
classification results}

\item{threshold}{Integer; Where classes containing a number of cells equal to
or less than the threshold will be grouped with the most similar class.}
}
\description{
Calculates Euclidean distance in PCA space.
}
\author{
Jason T. Serviss
}
