% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\docType{methods}
\name{run.tsne}
\alias{run.tsne}
\alias{run.tsne,spCounts-method}
\title{Run Tsne}
\usage{
run.tsne(x, ...)

\S4method{run.tsne}{spCounts}(x, plot.callback, k = 3, max_iter = 5000,
  perplexity = 30, initial_dims = 50, ...)
}
\arguments{
\item{...}{Additional arguments to pass on}

\item{counts}{Counts object.}
}
\value{
Ercc fraction plot.
}
\description{
Subtitle
}
\details{
Imports count and count.ercc data to a sp.scRNAseq object.
}
\examples{

#use demo data

#run function

}
\author{
Jason T. Serviss
}
\keyword{run.tsne}

