% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\name{runTsne}
\alias{runTsne}
\alias{runTsne}
\title{runTsne}
\usage{
runTsne(my.dist, dims = 2, theta = 0, initial_dims = 50,
  max_iter = 2000, perplexity = 10, seed = 11, is_distance = TRUE, ...)
}
\arguments{
\item{my.dist}{A distance object typically produced with pearsonsDist.}

\item{dims}{Argument to Rtsne. Numeric indicating the output dimensions.}

\item{theta}{Argument to
[Rtsne](https://cran.r-project.org/web/packages/Rtsne/index.html).}

\item{initial_dims}{Argument to
[Rtsne](https://cran.r-project.org/web/packages/Rtsne/index.html).}

\item{max_iter}{Argument to
[Rtsne](https://cran.r-project.org/web/packages/Rtsne/index.html).}

\item{perplexity}{Argument to
[Rtsne](https://cran.r-project.org/web/packages/Rtsne/index.html).}

\item{seed}{The desired seed to set before running.}

\item{is_distance}{Argument to
[Rtsne](https://cran.r-project.org/web/packages/Rtsne/index.html).}

\item{...}{Additional arguments to pass on}
}
\value{
A matrix containing the mean value for each gene for each
   classification group.
}
\description{
Calculates the x and y coordinates of the mean of each classif	ied group.
}
\details{
This method is typically only used in conjunction with plotting. It
calculates the 2 dimensional location of the mean of each classified group
in the supplied unsupervised dimensionality reduction (t-SNE) data
representation.
}
\examples{

s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[ ,s], testErcc[, s])
select <- spTopMax(cObjSng, 10)
my.dist <- pearsonsDist(cObjSng, select)
tsne <- runTsne(my.dist, max_iter = 10)


}
\author{
Jason T. Serviss
}
\keyword{runTsne}
