% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleSingletsArma}
\alias{sampleSingletsArma}
\title{sampleSingletsArma}
\usage{
sampleSingletsArma(classes)
}
\arguments{
\item{classes}{Character; a character vector of classes with length equal to
the number of cells for which counts exist.}
}
\description{
This function takes a character vector of classes/cell with the same order
as the cells in the counts matrix. It returns one random index per unique
class and returns them as a integer vector.
}
\author{
Jason T. Serviss
}
