% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\name{selectClustersOnEdge}
\alias{selectClustersOnEdge}
\alias{selectClustersOnEdge}
\title{selectClustersOnEdge}
\usage{
selectClustersOnEdge(spSwarm, edge.cutoff, clust1, clust2)
}
\arguments{
\item{spSwarm}{An spSwarm object.}

\item{edge.cutoff}{The minimum fraction to consider (?).}

\item{clust1}{A character vector of length 1 indicating one node which forms
an edge with clust2 for which multiplets contributing to that edge will be
reported.}

\item{clust2}{A character vector of length 1 indicating one node which forms
an edge with clust1 for which multiplets contributing to that edge will be
reported.}

\item{object}{spRSwarm object.}

\item{.Object}{Internal object.}

\item{...}{additional arguments to pass on}
}
\value{
spSwarm connection strengths and p-values.
}
\description{
Subtitle
}
\details{
Description
}
\examples{

#use demo data


#run function


}
\author{
Jason T. Serviss
}
\keyword{selectClustersOnEdge}
