% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{selectTopVar}
\alias{selectTopVar}
\title{selectTopVar}
\usage{
selectTopVar(cpm, n)
}
\arguments{
\item{cpm}{matrix; Counts per million.}

\item{n}{Number of genes to select.}
}
\value{
A numeric vector containing the indices of selected genes.
}
\description{
Facilitates gene selection prior to unsupervised clustering.
}
\details{
Returns the index for the n genes (rows) with the maximum
variance in the spCounts object. The expression matrix in
the counts.cpm slot is used for the calculation.
}
\examples{
s <- selectTopVar(getData(CIMseqSinglets_test, "counts.cpm"), 10)
}
\author{
Jason T. Serviss
}
