% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.scRNAseq-package.R
\docType{package}
\name{sp.scRNAseq-package}
\alias{sp.scRNAseq}
\alias{sp.scRNAseq-package}
\title{A package for analyzing tissue spatiality in scRNAseq data.}
\description{
Description
}
\details{
\tabular{ll}{ Package: \tab sp.scRNAseq\cr Type: \tab Package\cr
Version: \tab 1.0\cr Date: \tab 2016-02-28\cr License: \tab GPL-3\cr }
}
\author{
Author: Jason T. Serviss
}
\references{
Reference to published application note (work in progress)
}
\keyword{package}

