% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\docType{methods}
\name{spAverageGroupExpression}
\alias{spAverageGroupExpression}
\alias{spAverageGroupExpression}
\alias{spAverageGroupExpression,spCounts-method}
\alias{spAverageGroupExpression,matrix-method}
\title{spAverageGroupExpression}
\usage{
spAverageGroupExpression(counts, ...)

\S4method{spAverageGroupExpression}{spCounts}(counts, classes, ...)

\S4method{spAverageGroupExpression}{matrix}(counts, classes, ...)
}
\arguments{
\item{counts}{A counts matrix or spCounts object.}

\item{...}{Additional arguments to pass on}

\item{classes}{Character; A character vector indicating the class of each input sample.}
}
\value{
Matrix; Contains mean values for each gene in each class.
}
\description{
Calculates the average expression value of each gene in each class.
}
\details{
This function calculates the mean expression of each gene in each class which
is necessary for downstream assignment of multuplets to specific singlet classes.
}
\examples{

#use demo data

#run function

}
\author{
Jason T. Serviss
}
\keyword{spAverageGroupExpression}

