% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\docType{methods}
\name{spMclust}
\alias{spMclust}
\alias{spMclust}
\alias{spMclust,matrix-method}
\title{spMclust}
\usage{
spMclust(unsupervised, ...)

\S4method{spMclust}{matrix}(unsupervised, seed, Gmax, ...)
}
\arguments{
\item{unsupervised}{Matrix; Contains matrix resulting from unsupervised clustering.}

\item{...}{Additional arguments to pass on}

\item{seed}{Integer; Used to set seed.}

\item{Gmax}{Integer; Indicates the maximum number of clusters for which BIC is
calculated. Passed to Mclust.}
}
\value{
Character; Contains classifications for each sample.
}
\description{
Classification of unsupervised clustering data.
}
\details{
This function calculates the mean expression of each gene in each class which
is necessary for downstream assignment of multuplets to specific singlet classes.
}
\examples{

#use demo data

#run function

}
\author{
Jason T. Serviss
}
\keyword{spMclust}

