% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\docType{methods}
\name{spNtopMax}
\alias{spNtopMax}
\alias{spNtopMax}
\alias{spNtopMax,spCounts-method}
\alias{spNtopMax,matrix-method}
\title{spNtopMax}
\usage{
spNtopMax(counts, ...)

\S4method{spNtopMax}{spCounts}(counts, n, ...)

\S4method{spNtopMax}{matrix}(counts, n, ...)
}
\arguments{
\item{counts}{A counts matrix or spCounts object.}

\item{...}{Additional arguments to pass on}

\item{n}{Integer; the number of gene indices to return.}
}
\value{
Numeric; Indices of selected genes.
}
\description{
Subtitle
}
\details{
Ranks genes by their expression and returns the index
for the number of genes with maximum expression according
to the value of \emph{n}.
}
\examples{

#use demo data

#run function

}
\author{
Jason T. Serviss
}
\keyword{spNtopMax}

