% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\docType{methods}
\name{spNtopVar}
\alias{spNtopVar}
\alias{spNtopVar}
\alias{spNtopVar,spCounts-method}
\alias{spNtopVar,matrix-method}
\title{spNtopVar}
\usage{
spNtopVar(counts, ...)

\S4method{spNtopVar}{spCounts}(counts, n, ...)

\S4method{spNtopVar}{matrix}(counts, n, ...)
}
\arguments{
\item{counts}{A counts matrix or spCounts object.}

\item{...}{Additional arguments to pass on}

\item{n}{Integer; the number of gene indices to return.}
}
\value{
Numeric; Indices of selected genes.
}
\description{
Subtitle
}
\details{
Ranks genes by their variance and returns the index
for the number of genes with maximum variance according
to the value of \emph{n}.
}
\examples{

#use demo data

#run function

}
\author{
Jason T. Serviss
}
\keyword{spNtopVar}

