% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\docType{methods}
\name{spPearsonDist}
\alias{spPearsonDist}
\alias{spPearsonDist}
\alias{spPearsonDist,spCounts-method}
\alias{spPearsonDist,matrix-method}
\title{spPearsonDist}
\usage{
spPearsonDist(counts, ...)

\S4method{spPearsonDist}{spCounts}(counts, selectIdx, ...)

\S4method{spPearsonDist}{matrix}(counts, selectIdx, ...)
}
\arguments{
\item{counts}{A counts matrix or spCounts object.}

\item{...}{Additional arguments to pass on}

\item{selectIdx}{Numeric; A vector indicating the indices of genes to use.}
}
\value{
Dist; containg 1-Pearson's correlation of samples.
}
\description{
Subtitle
}
\details{
Calculates 1-Pearson's correlation of samples using
all genes specified by \emph{selectIdx}.
}
\examples{

#use demo data

#run function

}
\author{
Jason T. Serviss
}
\keyword{spPearsonDist}

