% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlot.R
\docType{methods}
\name{spPlot}
\alias{spPlot}
\alias{spPlot}
\alias{spPlot,spCounts-method}
\alias{spPlot,spUnsupervised-method}
\alias{spPlot,spSwarm-method}
\title{spPlot}
\usage{
spPlot(x, ...)

\S4method{spPlot}{spCounts}(x, y, type, markers = NULL, ...)

\S4method{spPlot}{spUnsupervised}(x, y = NULL, type, markers = NULL, ...)

\S4method{spPlot}{spSwarm}(x, y, z = NULL, layout = "tsne", loop = FALSE,
  markers = NULL, edge.cutoff = 0, min.pval = 0.05, min.num.edges = 1,
  ...)
}
\arguments{
\item{x}{An spCounts, spUnsupervised, or spSwarm object.}

\item{...}{additional arguments to pass on}

\item{y}{An spCounts, spUnsupervised, or spSwarm object.}

\item{type}{Can be "ercc", "markers", or .......}

\item{markers}{A character vector with 2 markers to plot.}

\item{z}{An spCounts object.}

\item{layout}{Specify "tsne" for overlaying connections over the tSNE or igraph for graph layout.}

\item{loop}{Logical; TRUE if "self" connections should be turned on.}
}
\value{
The spPlot function returns an object of class spCounts.
}
\description{
Subtitle
}
\details{
Imports count and count.ercc data to a sp.scRNAseq object.
}
\examples{

#use demo data
data(expData)

#run function

}
\author{
Jason T. Serviss
}
\keyword{spPlot}
