% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-classes.R, R/initialize-methods.R, R/show-methods.R, R/spSwarm.R
\docType{class}
\name{spSwarm-class}
\alias{.spSwarm}
\alias{getData,spSwarm-method}
\alias{initialize,spSwarm-method}
\alias{show,spSwarm-method}
\alias{spSwarm}
\alias{spSwarm,spUnsupervised-method}
\alias{spSwarm-class}
\title{spSwarm}
\usage{
\S4method{getData}{spSwarm}(x, n = NULL)

\S4method{initialize}{spSwarm}(.Object, ..., spSwarm, codedSwarm,
  spUnsupervised, arguments)

\S4method{show}{spSwarm}(object)

spSwarm(spUnsupervised, ...)

\S4method{spSwarm}{spUnsupervised}(spUnsupervised, limit = "none",
  maxiter = 10, swarmsize = 150, minstep = 1e-16, minfunc = 1e-16,
  cutoff = 0.2, cores = 1, ...)
}
\arguments{
\item{x}{A spSwarm object.}

\item{n}{Data to extract from spSwarm object.}

\item{.Object}{Internal object.}

\item{...}{additional arguments to pass on}

\item{spSwarm}{The spSwarm results.}

\item{codedSwarm}{The spSwarm results after the cutoff has been applied.}

\item{spUnsupervised}{An spUnsupervised object.}

\item{arguments}{Argumetns passed to the spSwarm function.}

\item{object}{spSwarm object.}

\item{limit}{Randomly select a limited number of samples to perform swarm optimization on.}

\item{maxiter}{pySwarm argument indicating the maximum optimization iterations.}

\item{swarmsize}{pySwarm argument indicating the number of particals in the swarm.}

\item{minstep}{pySwarm argument indicating the stepsize of swarm’s best position before search termination.}

\item{minfunc}{pySwarm argument indicating the minimum change of swarm’s best objective value before search termination.}

\item{cutoff}{The cutoff used to generate codedSwarm.}

\item{cores}{The number of cores to be used while running spSwarm.}
}
\value{
spSwarm output.
}
\description{
Subtitle
}
\details{
Description
}
\examples{

#use demo data


#run function


}
\author{
Jason T. Serviss
}
\keyword{pySwarm}
\keyword{spSwarm,}

