% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMseqSwarm.R
\name{spSwarmPoisson}
\alias{spSwarmPoisson}
\title{spSwarmPoisson}
\usage{
spSwarmPoisson(swarm, edge.cutoff = 0, min.pval = 1,
  min.num.edges = 0, ...)
}
\arguments{
\item{swarm}{A CIMseqSwarm object.}

\item{edge.cutoff}{numeric; The minimum fraction to consider (?).}

\item{min.pval}{numeric; Minimum p-value to report.}

\item{min.num.edges}{numeric; Minimum number of observed edges to report a 
connection.}

\item{...}{additional arguments to pass on}
}
\value{
CIMseqSwarm connection weights and p-values.
}
\description{
Subtitle
}
\details{
Description
}
\examples{

#use demo data
output <- spSwarmPoisson(CIMseqSwarm_test, 1/10.5)


}
\author{
Jason T. Serviss
}
