% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\name{spTopVar}
\alias{spTopVar}
\title{spTopVar}
\usage{
spTopVar(spCounts, n)
}
\arguments{
\item{spCounts}{An spCounts object.}

\item{n}{Number of genes to select.}
}
\value{
A numeric vector containing the indices of selected genes.
}
\description{
Facilitates gene selection prior to unsupervised clustering.
}
\details{
Returns the index for the n genes (rows) with the maximum
variance in the spCounts object. The expression matrix in
the counts.cpm slot is used for the calculation.
}
\examples{

selected <- spTopVar(test_spCountsSng, 10)

}
\author{
Jason T. Serviss
}
\keyword{spTopVar}
