% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\docType{methods}
\name{spTsne}
\alias{spTsne}
\alias{spTsne}
\alias{spTsne,dist-method}
\title{spTsne}
\usage{
spTsne(distances, ...)

\S4method{spTsne}{dist}(distances, dims = 2, theta = 0, initial_dims = 50,
  max_iter = 2000, perplexity = 10, seed = 11, is_distance = TRUE, ...)
}
\arguments{
\item{distances}{Dist; potentially produced by \emph{spPearsonDist} function.}

\item{...}{Additional arguments to pass on}

\item{dims}{Integer; number of output dimensions. Passed to Rtsne.}

\item{theta}{Numeric; Speed/accuracy tradeoff. Value between 0 and 1, where
0 is the most correct and 1 is the fastest. Passed to Rtsne.}

\item{initial_dims}{Integer; Number of dims to be retained in the initial
PCA step. Passed to Rtsne.}

\item{max_iter}{Integer; Number of iterations. Passed to Rtsne.}

\item{perplexity}{Numeric; Perplexity parameter. Passed to Rtsne.}

\item{seed}{Integer; Used to set seed.}

\item{is_distance}{Logical; Currently should not be altered. Passed to Rtsne.}
}
\value{
Matrix; new representations for the objects.
}
\description{
Runs Rtsne function from Rtsne package.
}
\details{
This function can be used for unsupervised clustering of cells allowing
downstream cell classification based on the clusters formed.
}
\examples{

#use demo data
dist <- spPearsonDist(
     testCounts[ ,grepl("s.", colnames(testCounts))],
     selectIdx=1:nrow(testCounts)
)

#run function
tsne <- spTsne(dist, max_iter=2, theta=1)

}
\author{
Jason T. Serviss
}
\keyword{spTsne}

