% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-classes.R, R/initialize-methods.R,
%   R/show-methods.R, R/spPlot.R, R/spUnsupervised.R
\docType{class}
\name{spUnsupervised-class}
\alias{spUnsupervised-class}
\alias{.spUnsupervised}
\alias{getData,spUnsupervised-method}
\alias{unsupervisedC}
\alias{unsupervisedC,spUnsupervised-method}
\alias{groupMeans}
\alias{groupMeans,spUnsupervised-method}
\alias{classification}
\alias{classification,spUnsupervised-method}
\alias{selectIdx}
\alias{selectIdx,spUnsupervised-method}
\alias{unsupervisedC<-}
\alias{unsupervisedC<-,spUnsupervised-method}
\alias{groupMeans<-}
\alias{groupMeans<-,spUnsupervised-method}
\alias{classification<-}
\alias{classification<-,spUnsupervised-method}
\alias{selectIdx<-}
\alias{selectIdx<-,spUnsupervised-method}
\alias{initialize,spUnsupervised-method}
\alias{show,spUnsupervised-method}
\alias{plot,spCounts,spUnsupervised-method}
\alias{spUnsupervised}
\alias{spUnsupervised}
\alias{spUnsupervised,missing-method}
\alias{spUnsupervised,spCounts-method}
\title{spUnsupervised}
\usage{
\S4method{getData}{spUnsupervised}(object, n = NULL)

unsupervisedC(object)

\S4method{unsupervisedC}{spUnsupervised}(object)

groupMeans(object)

\S4method{groupMeans}{spUnsupervised}(object)

classification(object)

\S4method{classification}{spUnsupervised}(object)

selectIdx(object)

\S4method{selectIdx}{spUnsupervised}(object)

unsupervisedC(object) <- value

\S4method{unsupervisedC}{spUnsupervised}(object) <- value

groupMeans(object) <- value

\S4method{groupMeans}{spUnsupervised}(object) <- value

classification(object) <- value

\S4method{classification}{spUnsupervised}(object) <- value

selectIdx(object) <- value

\S4method{selectIdx}{spUnsupervised}(object) <- value

\S4method{initialize}{spUnsupervised}(.Object, ..., unsupervisedC, groupMeans,
  classification, selectIdx)

\S4method{show}{spUnsupervised}(object)

\S4method{plot}{spCounts,spUnsupervised}(x, y, type, markers = NULL, ...)

spUnsupervised(spCounts, ...)

\S4method{spUnsupervised}{missing}(spCounts, ...)

\S4method{spUnsupervised}{spCounts}(spCounts, theta = 0, dims = 2,
  max_iter = 20000, perplexity = 10, initial_dims = 50, Gmax = 50,
  seed = 11, geneSelectType = "max", max = 2000, genes = NULL, ...)
}
\arguments{
\item{object}{spUnsupervised object.}

\item{n}{Data to extract from spUnsupervised object.}

\item{value}{Data to replace in spUnsupervised object.}

\item{.Object}{Internal object.}

\item{...}{Additional arguments to pass on}

\item{unsupervisedC}{tSNE results.}

\item{groupMeans}{The mean gene expression values for each cell type (classificaiton).}

\item{classification}{Post-tSNE cell type classification. Typically determined by the mclust package.}

\item{selectIdx}{The indexes of the genes picked for use in spUnsupervised. Used in spSwarm.}

\item{x}{Default plot param, an spCounts object containing singlets.}

\item{y}{Default plot param, an spCounts object containing multuplets.}

\item{type}{Character; The type of plot desired. Currently \emph{markers} or \emph{ercc}.}

\item{markers}{Markers/genes to plot.}

\item{spCounts}{spCounts object.}

\item{theta}{Passed to Rtsne.}

\item{dims}{The dimensions of the resulting tsne. Passed to tsne function.}

\item{max_iter}{The max number of tSNE iterations. Passed to tsne function.}

\item{perplexity}{The perplexity argument to tsne. Passed to tsne function.}

\item{initial_dims}{The initial dimensions argument. Passed to tsne function.}

\item{Gmax}{A numeric vector of 1:Gmax passed as the "G" argument to Mclust.}

\item{seed}{Sets the seed before running tSNE.}

\item{geneSelectType}{Decides if genes included are picked by their maximum expression
or maximum variance. Can be either "max", "var", or "manual". If "manual" the
genes argument must also be specified.}

\item{max}{The max number of genes to include based either on maximum expression
or maximum variance which is decided by the "type" paramater.}

\item{genes}{If type = manual, genes to be included are specified here as a character
vector. These must match the rownames in the counts variable.}
}
\value{
spUnsupervised object.
}
\description{
Subtitle
}
\details{
Description.
}
\examples{

#use demo data

#run function

}
\author{
Jason T. Serviss
}
\keyword{spUnsupervised}

