% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subsetSingletsEigen}
\alias{subsetSingletsEigen}
\title{subsetSingletsEigen}
\usage{
subsetSingletsEigen(singlets, idxToSubset)
}
\arguments{
\item{singlets}{Matrix; a counts matrix with cells/samples as columns and 
genes as rows.}

\item{idxToSubset}{Integer; the indexes of cells/samples to be subset. 
Typically generared using the \code{\link{subsetSingletsEigen}} function.}
}
\description{
This function takes a counts matrix and subsets the columns according to the
indices provided by the idxToSubset argument.
}
\author{
Jason T. Serviss
}
