% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psoptim.R
\name{swarmInit}
\alias{swarmInit}
\title{swarmInit}
\usage{
swarmInit(singlets, k, null.weight = 1, seed = 9238232)
}
\arguments{
\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{k}{integer; Number of cells in each combination, i.e. 2 gives doublets, 
3 gives triplets, etc. Must be >= 2.}

\item{null.weight}{numeric; Weight to adjust the amount of noise in the 
initialization positions. Default = 1.}

\item{seed}{integer; A seed to set.}
}
\value{
A matrix with the swarm initialization state.
}
\description{
Helper function to calculate the swarm initialization positions to provide
as the swarmInit argument to CIMseqSwarm.
}
\author{
Martin Enge
}
