% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{test_spCountsMul}
\alias{test_spCountsMul}
\title{spCounts object with multiplets corresponding to test dataset.}
\format{spCounts
\describe{
    \item{counts}{See \code{\link{spCounts}}.}
    \item{counts.log}{See \code{\link{spCounts}}.}
    \item{counts.cpm}{See \code{\link{spCounts}}.}
    \item{counts.ercc}{See \code{\link{spCounts}}.}
}}
\usage{
test_spCountsMul
}
\value{
An spCounts object.
}
\description{
spCounts object with multiplets.
}
\examples{
data(test_spCountsMul)

}
