% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{test_spCountsSng}
\alias{test_spCountsSng}
\title{spCounts object with singlets corresponding to test dataset.}
\format{spCounts
\describe{
    \item{counts}{See \code{\link{spCounts}}.}
    \item{counts.log}{See \code{\link{spCounts}}.}
    \item{counts.cpm}{See \code{\link{spCounts}}.}
    \item{counts.ercc}{See \code{\link{spCounts}}.}
}}
\usage{
test_spCountsSng
}
\value{
An spCounts object.
}
\description{
spCounts object with singlets.
}
\examples{
data(test_spCountsSng)

}
