% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{test_spSwarm}
\alias{test_spSwarm}
\title{spSwarm object corresponding to test dataset.}
\format{spSwarm
\describe{
 \item{spSwarm}{See \code{\link{spSwarm}}.}
 \item{costs}{See \code{\link{spSwarm}}.}
 \item{convergence}{See \code{\link{spSwarm}}.}
 \item{singletIdx}{See \code{\link{spSwarm}}.}
 \item{arguments}{See \code{\link{spSwarm}}.}
}}
\usage{
test_spSwarm
}
\value{
spSwarm object.
}
\description{
spSwarm object.
}
\examples{
data(test_spSwarm)

}
