% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{test_spUnsupervised}
\alias{test_spUnsupervised}
\title{spUnsupervised object corresponding to test dataset.}
\format{spUnsupervised
\describe{
    \item{tsne}{See \code{\link{spUnsupervised}}.}
    \item{tsneMeans}{See \code{\link{spUnsupervised}}.}
    \item{classification}{See \code{\link{spUnsupervised}}.}
    \item{uncertainty}{See \code{\link{spUnsupervised}}.}
    \item{selectInd}{See \code{\link{spUnsupervised}}.}
}}
\usage{
test_spUnsupervised
}
\value{
An spUnsupervised object.
}
\description{
spUnsupervised object.
}
\examples{
data(test_spUnsupervised)

}
