% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCounts.R
\docType{data}
\name{testCounts}
\alias{testCounts}
\title{Simulated data used to demonstrate sp.scRNAseq package.}
\format{Matrix
\describe{
\item{rownames}{Gene counts}
\item{colnames}{Cells}
}}
\usage{
testCounts
}
\value{
matrix
}
\description{
Test counts.
}
\examples{
testCounts

}

