% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testErcc.R
\docType{data}
\name{testErcc}
\alias{testErcc}
\title{Simulated data used to demonstrate sp.scRNAseq package.}
\format{Matrix
\describe{
\item{rownames}{ERCC spike-in counts}
\item{colnames}{Cells}
}}
\usage{
testErcc
}
\value{
matrix
}
\description{
Test ercc.
}
\examples{
testErcc

}
