% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{testMeta}
\alias{testMeta}
\title{Meta data for the testCounts dataset.}
\format{tibble
\describe{
    \item{sample}{Sample ID}
    \item{cellNumber}{Indicates if samples are singlets or multiplets.}
    \item{cellTypes}{Indicates the cell types included in the sample.}
}}
\usage{
testMeta
}
\value{
Tibble with meta data.
}
\description{
Test counts meta data.
}
\examples{
data(testMeta)

}
