% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSwa.R
\docType{data}
\name{testSwa}
\alias{testSwa}
\title{spSwarm object corresponding to test dataset.}
\format{spUnsupervised
\describe{
    \item{spSwarm}{See \code{\link{spSwarm}}.}
    \item{costs}{See \code{\link{spSwarm}}.}
    \item{convergence}{See \code{\link{spSwarm}}.}
    item{arguments}{See \code{\link{spSwarm}}.}
}}
\usage{
testSwa
}
\value{
spSwarm object.
}
\description{
spSwarm object.
}
\examples{
data(testData)

}
