% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testUns.R
\docType{data}
\name{testUns}
\alias{testUns}
\title{spUnsupervised object corresponding to test dataset.}
\format{spUnsupervised
\describe{
    \item{tsne}{See \code{\link{spUnsupervised}}.}
    \item{tsneMeans}{See \code{\link{spUnsupervised}}.}
    \item{groupMeans}{See \code{\link{spUnsupervised}}.}
    \item{classification}{See \code{\link{spUnsupervised}}.}
    \item{uncertainty}{See \code{\link{spUnsupervised}}.}
    \item{selectInd}{See \code{\link{spUnsupervised}}.}
}}
\usage{
testUns
}
\value{
An spUnsupervised object.
}
\description{
spUnsupervised object.
}
\examples{
data(testData)

}
