% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyFunctions.R
\name{tidySinglets}
\alias{tidySinglets}
\title{tidySinglets}
\usage{
tidySinglets(singlets)
}
\arguments{
\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}
}
\description{
Tidy CIMseqSinglets objects. Drops all associated counts data.
}
\examples{

tidySinglets(CIMseqSinglets_test)

}
\author{
Jason T. Serviss
}
\keyword{tidySinglets}
