% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyFunctions.R
\name{tidySwarm}
\alias{tidySwarm}
\title{tidySwarm}
\usage{
tidySwarm(swarm)
}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}
}
\description{
Tidy CIMseqSwarm objects.
}
\examples{

tidySwarm(CIMseqSwarm_test)

}
\author{
Jason T. Serviss
}
\keyword{tidySwarm}
