% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\name{tsneGroupMeans}
\alias{tsneGroupMeans}
\title{tsneGroupMeans}
\usage{
tsneGroupMeans(data, classes)
}
\arguments{
\item{data}{Singlet 2D tsne.}

\item{classes}{A character vector indicating the class of each singlet.}
}
\value{
A matrix containing the mean value for each gene for each
   classification group.
}
\description{
Calculates the x and y coordinates of the mean of each classified group.
}
\details{
This method is typically only used in conjunction with plotting. It
calculates the 2 dimensional location of the mean of each classified group
in the supplied unsupervised dimensionality reduction (t-SNE) data
representation.
}
\examples{

meanGroupPos <- tsneGroupMeans(
  getData(test_spUnsupervised, "tsne"),
  getData(test_spUnsupervised, "classification")
)

}
\author{
Jason T. Serviss
}
\keyword{tsneGroupMeans}
