% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vecToMat}
\alias{vecToMat}
\title{vecToMat}
\usage{
vecToMat(vec, nr, nc)
}
\arguments{
\item{vec}{Numeric; the vector to reformat.}

\item{nr}{integer; length 1 integer indicating the number of matrix rows.}

\item{nc}{integer; length 1 integer indicating the number of matrix columns.}
}
\description{
This function takes a vector and reformats it to a matrix in a column-wise
fashion. Since Armadillo "builds" the matrix column-wise, some code tricks
are employed to automagically deliever the expected result.
}
\author{
Jason T. Serviss
}
