% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vecToMatArma}
\alias{vecToMatArma}
\title{vecToMatArma}
\usage{
vecToMatArma(vec, nr, nc)
}
\arguments{
\item{vec}{Numeric; the vector to reformat.}

\item{nr}{integer; length 1 integer indicating the number of matrix rows.}

\item{nc}{integer; length 1 integer indicating the number of matrix columns.}
}
\description{
This function takes a vector and reformats it to a matrix in a column-wise
fashion.
}
\author{
Jason T. Serviss
}
