% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMseqSwarm.R
\name{adjustFractions}
\alias{adjustFractions}
\title{adjustFractions}
\usage{
adjustFractions(singlets, multiplets, swarm, binary = TRUE,
  theoretical.max = Inf, ...)
}
\arguments{
\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}

\item{swarm}{CIMseqSwarm or matrix; A CIMseqSwarm object or a matrix of 
fractions.}

\item{binary}{logical; Indicates if adjusted fractions should be returned as
binary values.}

\item{theoretical.max}{integer; See \code{\link{estimateCells}}.}

\item{...}{additional arguments to pass on}
}
\value{
Adjusted fractions matrix.
}
\description{
Transforms continious fractions into a binary martix where 1 indicates a
connection between classes.
}
\examples{

#use demo data


}
\author{
Jason T. Serviss
}
