% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\name{appropriateSinglets}
\alias{appropriateSinglets}
\title{appropriateSinglets}
\usage{
appropriateSinglets(spUnsupervised, spCountsSng, idx, selectInd)
}
\arguments{
\item{spUnsupervised}{An spUnsupervised object.}

\item{spCountsSng}{An spCounts object with singlets.}

\item{idx}{Singlet indices to subset. Generated with the sampleSinglets 
function.}

\item{selectInd}{Indices of selected features for deconvolution.}

\item{...}{additional arguments to pass on}
}
\value{
Appropriated singlets.
}
\description{
Subtitle
}
\details{
Description
}
\examples{

#use demo data


}
\author{
Jason T. Serviss
}
\keyword{appropriateSinglets}
