% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\name{averageGroupExpression}
\alias{averageGroupExpression}
\alias{averageGroupExpression}
\title{averageGroupExpression}
\usage{
averageGroupExpression(data, classes, weighted, uncertainty = NULL)
}
\arguments{
\item{data}{Singlet expression matrix.}

\item{classes}{A character vector indicating the class of each singlet.}

\item{weighted}{Logical indicating if the group means shoule be weighted with
the uncertainty.}

\item{uncertainty}{A numeric vector indicating the uncertainty if weighted is
TRUE.}
}
\value{
A matrix containing the mean value for each gene for each
classification group.
}
\description{
This output from this function is utilized to represent each group classified
group during swarm optimization.
}
\details{
Calculate the average expression of each gene within each classification
group. Note that typically only singlets are input to this method.
}
\examples{

s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])
select <- spTopMax(cObjSng, 10)
my.dist <- pearsonsDist(cObjSng, select)
tsne <- runTsne(my.dist, max_iter = 2000)
mclustOUT <- runMclust(tsne, 7, 11)
class <- mclustOUT[[1]]
uncertainty <- mclustOUT[[2]]

#Weighted mean
weighted <- TRUE
averageExp <- averageGroupExpression(cObjSng, class, weighted, uncertainty)

#Unweighted mean
weighted <- FALSE
averageExp <- averageGroupExpression(cObjSng, class, weighted)

}
\author{
Jason T. Serviss
}
\keyword{averageGroupExpression}
