% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\name{calcResiduals}
\alias{calcResiduals}
\alias{calcResiduals}
\title{calcResiduals}
\usage{
calcResiduals(spCounts, spUnsupervised, spSwarm, clusters = NULL,
  edge.cutoff = NULL, distFun = function(fractions, cellTypes, oneMultiplet,
  e, cellNumber, ...) {     if (sum(fractions) == 0) {         return(999999999)
      }     normFractions <- fractions/sum(fractions)     cellTypes <-
  cellTypes/mean(cellTypes)     a <- .makeSyntheticSlice(cellTypes,
  normFractions)     a <- a/mean(a)     k <- length(which(normFractions > 0))   
   penalty <- .complexityPenilty(k, e, cellNumber)     abs((oneMultiplet - a)/(a
  + 1)) * penalty }, ...)
}
\arguments{
\item{spCounts}{An spCounts object with multiplets.}

\item{spUnsupervised}{An spUnsupervised object.}

\item{spSwarm}{An spSwarm object.}

\item{clusters}{A character vector of length 2 indicating 2 classes to
specifically extract residuals from.}

\item{edge.cutoff}{The minimum fraction to consider (?).}

\item{distFun}{The distance function to be used to calculate the residuals.}

\item{...}{additional arguments to pass on}
}
\value{
spSwarm connection strengths and p-values.
}
\description{
Subtitle
}
\details{
Calculates the residuals for each gene and multiplet after deconvolution
based on the spSwarm results.
}
\examples{

#s <- grepl("^s", colnames(testCounts))
#cObjMul <- spCounts(testCounts[, !s], testErcc[, !s])
#output <- calcResiduals(cObjMul, testUns, testSwa)

}
\author{
Jason T. Serviss
}
\keyword{calcResiduals}
