% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\name{calcResiduals}
\alias{calcResiduals}
\alias{calcResiduals}
\title{calcResiduals}
\usage{
calcResiduals(spCountsMul, spSwarm, ...)
}
\arguments{
\item{spCountsMul}{An spCounts object with multiplets.}

\item{spSwarm}{An spSwarm object.}

\item{...}{additional arguments to pass on}
}
\value{
spSwarm connection strengths and p-values.
}
\description{
Subtitle
}
\details{
Calculates the residuals for each gene and multiplet after deconvolution
based on the spSwarm results.
}
\author{
Jason T. Serviss
}
\keyword{calcResiduals}
