% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\name{calcResiduals}
\alias{calcResiduals}
\alias{calcResiduals}
\title{calcResiduals}
\usage{
calcResiduals(spCounts, spUnsupervised, spSwarm, clusters = NULL,
  edge.cutoff = NULL, distFun = function(frac, multiplets) {     abs(frac -
  multiplets) }, ...)
}
\arguments{
\item{spCounts}{An spCounts object with multiplets.}

\item{spUnsupervised}{An spUnsupervised object.}

\item{spSwarm}{An spSwarm object.}

\item{clusters}{A character vector of length 2 indicating 2 classes to
specifically extract residuals from.}

\item{edge.cutoff}{The minimum fraction to consider (?).}

\item{...}{additional arguments to pass on}

\item{object}{spRSwarm object.}

\item{.Object}{Internal object.}
}
\value{
spSwarm connection strengths and p-values.
}
\description{
Subtitle
}
\details{
Calculates the residuals for each gene and multiplet after deconvolution
based on the spSwarm results.
}
\examples{

#use demo data


#run function


}
\author{
Jason T. Serviss
}
\keyword{calcResiduals}
