% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\name{calcResiduals}
\alias{calcResiduals}
\alias{calcResiduals}
\title{calcResiduals}
\usage{
calcResiduals(spCounts, spUnsupervised, spSwarm, clusters = NULL,
  edge.cutoff = NULL, distFun = function(frac, multiplets) {    
  (abs(multiplets - frac)/(frac + 1)) }, ...)
}
\arguments{
\item{spCounts}{An spCounts object with multiplets.}

\item{spUnsupervised}{An spUnsupervised object.}

\item{spSwarm}{An spSwarm object.}

\item{clusters}{A character vector of length 2 indicating 2 classes to
specifically extract residuals from.}

\item{edge.cutoff}{The minimum fraction to consider (?).}

\item{distFun}{The distance function to be used to calculate the residuals.}

\item{...}{additional arguments to pass on}
}
\value{
spSwarm connection strengths and p-values.
}
\description{
Subtitle
}
\details{
Calculates the residuals for each gene and multiplet after deconvolution
based on the spSwarm results.
}
\examples{

s <- grepl("^s", colnames(testCounts))
cObjMul <- spCounts(testCounts[, !s], testErcc[, !s])
output <- calcResiduals(cObjMul, testUns, testSwa)

#run function


}
\author{
Jason T. Serviss
}
\keyword{calcResiduals}
