% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/spSwarm.R
\name{calculateCost}
\alias{calculateCost}
\alias{calculateCost}
\alias{calculateCost}
\title{calculateCost}
\usage{
calculateCost(spCountsMul, spUnsupervised, fractions, distFun, e = NULL,
  cellNumbers = NULL, ...)

calculateCost(spCountsMul, spUnsupervised, fractions, distFun, e = NULL,
  cellNumbers = NULL, ...)
}
\arguments{
\item{spCountsMul}{An spCounts object containing multiplets.}

\item{spUnsupervised}{An spUnsupervised object.}

\item{fractions}{Numeric; a numeric vector with length equal to
ncol(singlets) indicating the fractions that each column should be
multiplied with.}

\item{distFun}{The distance function to use for the cost calculation.}

\item{e}{Argument to dtsnCellNum distance function.}

\item{cellNumbers}{Argument to dtsnCellNum distance function.}

\item{...}{additional arguments to pass on}

\item{oneMultiplet}{Integer; a integer vector of rounded counts per million
for one multiplet.}

\item{singletSubset}{Matrix; Numeric matrix with the preallocated singlets.
Each of the n synthetic multiplets should be stacked, i.e. rbind.}

\item{n}{Integer; length 1 integer indicating the number of synthetic
multiplets to generate.}

\item{fractions}{The fractions to calculate the cost with.}
}
\value{
The cost.
}
\description{
Wrapper for deconvolution cost function using the Armadillo C++ library.

Calculates the cost for a multiplet with a input of specific fractions.
}
\details{
Description
}
\examples{

#

}
\author{
Jason T. Serviss

Jason T. Serviss
}
\keyword{calculateCost}
