% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMseqSwarm.R
\docType{methods}
\name{calculateCosts}
\alias{calculateCosts}
\alias{calculateCosts,CIMseqSinglets,CIMseqMultiplets,CIMseqSwarm-method}
\title{calculateCosts}
\usage{
calculateCosts(singlets, multiplets, swarm, ...)


  \S4method{calculateCosts}{CIMseqSinglets,CIMseqMultiplets,CIMseqSwarm}(singlets,
  multiplets, swarm, fractions = NULL, ...)
}
\arguments{
\item{singlets}{A CIMseqSinglets object.}

\item{multiplets}{A CIMseqMultiplets object.}

\item{swarm}{A CIMseqSwarm object.}

\item{...}{additional arguments to pass on}

\item{fractions}{WILL PROBABLY BE REMOVED}
}
\value{
Costs
}
\description{
Description
}
\examples{

#

}
\author{
Jason T. Serviss
}
\keyword{calculateCosts}
