% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\docType{methods}
\name{calculateCosts}
\alias{calculateCosts}
\alias{calculateCosts}
\alias{calculateCosts,spCounts,spSwarm,numeric-method}
\title{calculateCosts}
\usage{
calculateCosts(spCountsMul, spSwarm, fractions, ...)

\S4method{calculateCosts}{spCounts,spSwarm,numeric}(spCountsMul, spSwarm,
  fractions, ...)
}
\arguments{
\item{spSwarm}{An spSwarm object.}

\item{...}{additional arguments to pass on}
}
\value{
Costs
}
\description{
Description
}
\examples{

#

}
\author{
Jason T. Serviss
}
\keyword{calculateCosts}
