% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMseqSwarm.R
\docType{methods}
\name{calculateCosts}
\alias{calculateCosts}
\alias{calculateCosts,CIMseqSinglets,CIMseqSwarm,numeric-method}
\title{calculateCosts}
\usage{
calculateCosts(multiplets, swarm, fractions, ...)

\S4method{calculateCosts}{CIMseqSinglets,CIMseqSwarm,numeric}(multiplets,
  swarm, fractions = NULL, ...)
}
\arguments{
\item{multiplets}{A CIMseqMultiplets object.}

\item{swarm}{A CIMseqSwarm object.}

\item{fractions}{WILL PROBABLY BE REMOVED}

\item{...}{additional arguments to pass on}
}
\value{
Costs
}
\description{
Description
}
\examples{

#

}
\author{
Jason T. Serviss
}
\keyword{calculateCosts}
