% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\docType{methods}
\name{changeCutoff}
\alias{changeCutoff}
\alias{changeCutoff,spSwarm-method}
\title{changeCutoff}
\usage{
changeCutoff(spSwarm, ...)

\S4method{changeCutoff}{spSwarm}(spSwarm, spCounts = NULL, error = NULL,
  cutoff, ...)
}
\arguments{
\item{spSwarm}{An spSwarm object.}

\item{...}{additional arguments to pass on}

\item{spCounts}{An spCounts object. Needed when counts.ercc should be used.}

\item{error}{Numeric; for use with counts.ercc. Lowers the calculated cutoff by error value.}

\item{cutoff}{The fraction below which a connection should not be considered.}
}
\value{
spSwarm output.
}
\description{
Subtitle
}
\details{
Description
}
\examples{

#use demo data


#run function


}
\author{
Jason T. Serviss
}
\keyword{changeCutoff}

