% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-classes.R, R/CIMseqSwarm.R,
%   R/initialize-methods.R, R/show-methods.R
\docType{class}
\name{CIMseqSwarm-class}
\alias{CIMseqSwarm-class}
\alias{getData,CIMseqSwarm-method}
\alias{c,CIMseqSwarm-method}
\alias{filterSwarm,CIMseqSwarm-method}
\alias{CIMseqSwarm}
\alias{CIMseqSwarm,CIMseqSinglets,CIMseqMultiplets-method}
\alias{initialize,CIMseqSwarm-method}
\alias{show,CIMseqSwarm-method}
\title{CIMseqSwarm}
\usage{
\S4method{getData}{CIMseqSwarm}(x, n = NULL)

\S4method{c}{CIMseqSwarm}(x, ...)

\S4method{filterSwarm}{CIMseqSwarm}(x, subset)

CIMseqSwarm(singlets, multiplets, ...)

\S4method{CIMseqSwarm}{CIMseqSinglets,CIMseqMultiplets}(singlets,
  multiplets, maxiter = 10, swarmsize = 150,
  nSyntheticMultiplets = 200, seed = 11, norm = TRUE,
  report = FALSE, reportRate = NA, permute = FALSE,
  singletIdx = NULL, swarmInit = NULL, psoControl = list(), ...)

\S4method{initialize}{CIMseqSwarm}(.Object, ..., fractions, costs,
  convergence, stats, singletIdx, arguments)

\S4method{show}{CIMseqSwarm}(object)
}
\arguments{
\item{x}{CIMseqSwarm; A CIMseqSwarm object.}

\item{n}{character; Data to extract from CIMseqSwarm object.}

\item{...}{additional arguments to pass on.}

\item{subset}{Indicates samples to be retained. Can be a character vector of 
sample names, a logical vector, or integer vector indicating sample indices.}

\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}

\item{maxiter}{integer; The maximum optimization iterations.}

\item{swarmsize}{integer; The number of swarm particals.}

\item{nSyntheticMultiplets}{Numeric; Value indicating the number of synthetic
multiplets to generate during deconvolution.}

\item{seed}{integer; Seed to set before running.}

\item{norm}{logical; Indicates if the sum of fractions should equal 1.}

\item{report}{logical; Indicates if additional reporting from the
optimization should be included.}

\item{reportRate}{integer; If report is TRUE, the iteration interval for 
which a report should be generated.}

\item{permute}{logical; indicates if genes should be permuted before
deconvolution. For use with permutation testing.
@param singletIdx list; Singlet indexes to be used to choose singlets and
synthesize synthetic multiplets. Facilitates using the same synthetic set 
e.g. with repeated runs or permutation.}

\item{singletIdx}{list; Indexes indicating singlets that were subset to 
synthesize synthetic multiplets. Facilitates recreation of the synthetic 
multiplets downstream.}

\item{swarmInit}{matrix; Initiation positions for the swarm.}

\item{psoControl}{list; Additional arguments to pso.2.0 (psoptim) function.}

\item{.Object}{Internal object.}

\item{fractions}{matrix; The deconvolution results.}

\item{costs}{numeric; The costs after optimization.}

\item{convergence}{character; The convergence output from \link[pso]{psoptim}.
One value per multiplet.}

\item{stats}{tbl_df; The stats output from \link[pso]{psoptim}}

\item{arguments}{list; Arguments passed to the CIMseqSwarm function.}

\item{object}{CIMseqSwarm; A CIMseqSwarm to show.}
}
\value{
CIMseqSwarm output.
}
\description{
Multiplet deconvolution.
}
\examples{

#use demo data

}
\author{
Jason T. Serviss
}
