% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotExtras.R
\name{convertToERCC}
\alias{convertToERCC}
\title{convertToERCC}
\usage{
convertToERCC(ercc, spCountsSng, spCountsMul)
}
\arguments{
\item{ercc}{The left axis values. Passes as ".".}

\item{spCountsSng}{spCounts; An spCounts object with singlets.}

\item{spCountsSng}{spCounts; An spCounts object with multiplets.}
}
\description{
A function to facilitate calculation of the second axis of the plotCounts
type "ercc" plot.
}
\author{
Jason T. Serviss
}
\keyword{convertToERCC}
