% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{countsPerClass}
\alias{countsPerClass}
\title{countsPerClass}
\usage{
countsPerClass(singlets)
}
\arguments{
\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}
}
\value{
A matrix containing the median value for each gene for each
 classification group.
}
\description{
Calculates median counts per class.
}
\examples{

output <- countsPerClass(CIMseqSinglets_test)

}
\author{
Jason T. Serviss
}
