% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{erccPerClass}
\alias{erccPerClass}
\title{erccPerClass}
\usage{
erccPerClass(singlets, multiplets)
}
\arguments{
\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}
}
\value{
A matrix containing the median value for each gene for each
 classification group.
}
\description{
Calculates median ercc reads per class.
}
\examples{

out <- erccPerClass(CIMseqSinglets_test, CIMseqMultiplets_test)

}
\author{
Jason T. Serviss
}
