% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\name{erccPerClass}
\alias{erccPerClass}
\alias{erccPerClass}
\title{erccPerClass}
\usage{
erccPerClass(spCountsSng, spCountsMul, spUnsupervised)
}
\arguments{
\item{spCountsSng}{An spCounts object containing singlets.}

\item{spCountsMul}{An spCounts object containing multiplets.}

\item{spUnsupervised}{An spUnsupervised object.}
}
\value{
A matrix containing the mean value for each gene for each
   classification group.
}
\description{
Calculates median ercc reads per class.
}
\examples{

#write something

}
\author{
Jason T. Serviss
}
\keyword{erccPerClass}
