% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCounts.R
\docType{methods}
\name{estimateCells}
\alias{estimateCells}
\alias{estimateCells,spCounts-method}
\title{estimateCells}
\usage{
\S4method{estimateCells}{spCounts}(spCountsSng, spCountsMul,
  warning = TRUE, ...)
}
\arguments{
\item{spCountsSng}{A spCounts object with singlets.}

\item{spCountsMul}{A spCounts object with multiplets.}

\item{warning}{logical; Indicates if a warning should be issued when all ERCC
counts for a sample are equal to 0. If this warning is issued it can be
satisfactorily resolved by setting the ERCC reads for these samples to NA.}

\item{...}{additional arguments to pass on}
}
\value{
A data frame including the fraction of ercc reads and cell counts for
   each sample.
}
\description{
Subtitle
}
\details{
Uses ERCC data to calculate the fraction of ERCC reads in the samples. In
addition, this function utilizes ERCC data to estimate the cell number
in each sample.
}
\examples{

output <- estimateCells(test_spCountsSng, test_spCountsMul)

}
\author{
Jason T. Serviss
}
\keyword{spCounts}
