% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCounts.R
\docType{methods}
\name{estimateCells}
\alias{estimateCells}
\alias{estimateCells,spCounts-method}
\title{estimateCells}
\usage{
\S4method{estimateCells}{spCounts}(spCountsSng, spCountsMul, ...)
}
\arguments{
\item{spCountsSng}{A spCounts object with singlets.}

\item{spCountsMul}{A spCounts object with multiplets.}

\item{...}{additional arguments to pass on}
}
\value{
A data frame including the fraction of ercc reads and cell counts for
   each sample.
}
\description{
Subtitle
}
\details{
Uses ERCC data to calculate the fraction of ERCC reads in the samples. In
addition, this function utilizes ERCC data to estimate the cell number
in each sample.
}
\examples{

#use test data
s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])
cObjMul <- spCounts(testCounts[, !s], testErcc[, !s])

#run function
output <- estimateCells(cObjSng, cObjMul)

}
\author{
Jason T. Serviss
}
\keyword{spCounts}
