% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{estimateTotalConnections}
\alias{estimateTotalConnections}
\title{estimateTotalConnections}
\usage{
estimateTotalConnections(singlets, multiplets)
}
\arguments{
\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}
}
\value{
The number of estimated connections.
}
\description{
Estimates the total number of connections in the multiplets.
}
\examples{

estimateTotalConnections(CIMseqSinglets_test, CIMseqMultiplets_test)

}
\author{
Jason T. Serviss
}
\keyword{estimateTotalConnections}
