% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{expectedInteractionFreq}
\alias{expectedInteractionFreq}
\title{expectedInteractionFreq}
\usage{
expectedInteractionFreq(singlets)
}
\arguments{
\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}
}
\value{
A table with the expected frequencies.
}
\description{
Calculates the expected interaction frequency between cell types.
}
\examples{

exfreq <- expectedInteractionFreq(CIMseqSinglets_test)

}
\author{
Jason T. Serviss
}
\keyword{expectedInteractionFreq}
