% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCounts.R
\docType{methods}
\name{filterCells}
\alias{filterCells}
\alias{filterCells,spCounts-method}
\title{Filter Cells}
\usage{
filterCells(x, ...)

\S4method{filterCells}{spCounts}(x, quantile.cut = 0.001,
  gene.name = "ACTB", ...)
}
\arguments{
\item{x}{Dataset of class Counts.}

\item{...}{additional arguments to pass on}

\item{quantile.cut}{"p" argument to qnorm.}

\item{gene.name}{Gene to filter on.}
}
\value{
The filterCells function returns an object of class Counts.
}
\description{
Select a cutoff for bad cells based on actin expression.
}
\details{
A method to filter low quality cells from the analysis.
}
\examples{

#use demo data

#run function

}
\author{
Jason T. Serviss
}
\keyword{filterCells}

