% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMseqSwarm.R
\docType{methods}
\name{getEdgesForMultiplet}
\alias{getEdgesForMultiplet}
\alias{getEdgesForMultiplet,CIMseqSwarm-method}
\title{getEdgesForMultiplet}
\usage{
getEdgesForMultiplet(swarm, ...)

\S4method{getEdgesForMultiplet}{CIMseqSwarm}(swarm, singlets, multiplets,
  multipletName = NULL, ...)
}
\arguments{
\item{swarm}{A CIMseqSwarm object.}

\item{...}{additional arguments to pass on}

\item{singlets}{A CIMseqSinglets object.}

\item{multiplets}{A CIMseqMultiplets object.}

\item{multipletName}{character; The name of the multiplet of interest.}
}
\value{
Edge names.
}
\description{
Returns the names of the edges detected in a multiplet.
}
\details{
Description
}
\examples{

output <- getEdgesForMultiplet(
CIMseqSwarm_test, CIMseqSinglets_test, CIMseqMultiplets_test,
"m.NJB00204.G04"
)

}
\author{
Jason T. Serviss
}
