% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\docType{methods}
\name{getEdgesForMultiplet}
\alias{getEdgesForMultiplet}
\alias{getEdgesForMultiplet}
\alias{getEdgesForMultiplet,spSwarm-method}
\title{getEdgesForMultiplet}
\usage{
getEdgesForMultiplet(spSwarm, ...)

\S4method{getEdgesForMultiplet}{spSwarm}(spSwarm, edge.cutoff, multiplet, ...)
}
\arguments{
\item{spSwarm}{An spSwarm object.}

\item{...}{additional arguments to pass on}

\item{edge.cutoff}{The minimum fraction to consider (?).}

\item{multiplet}{The name of the multiplet of interest.}
}
\value{
Edge names.
}
\description{
Returns the names of the edges are associated with a multiplet.
}
\details{
Description
}
\examples{

output <- getEdgesForMultiplet(testSwa, 1/10.5, "m.A1B1")

}
\author{
Jason T. Serviss
}
\keyword{getEdgesForMultiplet}
