% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMseqSwarm.R
\docType{methods}
\name{getMultipletsForEdge}
\alias{getMultipletsForEdge}
\alias{getMultipletsForEdge,CIMseqSwarm-method}
\title{getMultipletsForEdge}
\usage{
getMultipletsForEdge(swarm, ...)

\S4method{getMultipletsForEdge}{CIMseqSwarm}(swarm, singlets, multiplets,
  edges, theoretical.max = Inf, ...)
}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{...}{additional arguments to pass on}

\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}

\item{edges}{data.frame; Edges of interest. Edges are indicated by the nodes
they connect with one node in column one and the other node in column 2.}

\item{theoretical.max}{integer; See \code{\link{estimateCells}}.}
}
\value{
If the edges argument only includes one row, a vector of multiplet
   names is returned. If several edges are interogated a list is returned
   with one element per edge containing the names of the multiplets.
}
\description{
Returns the names of the multiplets that are associated with an edge.
}
\examples{

output <- getMultipletsForEdge(
CIMseqSwarm_test, 
CIMseqSinglets_test, 
CIMseqMultiplets_test, 
data.frame("A375", "HOS")
)

}
\author{
Jason T. Serviss
}
