% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMseqSwarm.R
\docType{methods}
\name{getMultipletsForEdge}
\alias{getMultipletsForEdge}
\alias{getMultipletsForEdge,CIMseqSwarm-method}
\title{getMultipletsForEdge}
\usage{
getMultipletsForEdge(swarm, ...)

\S4method{getMultipletsForEdge}{CIMseqSwarm}(swarm, edge.cutoff, edges,
  ...)
}
\arguments{
\item{swarm}{A CIMseqSwarm object.}

\item{...}{additional arguments to pass on}

\item{edge.cutoff}{numeric; The minimum fraction to consider (?).}

\item{edges}{data.frame; Edges of interest. Edges are indicated by the nodes
they connect with one node in column one and the other node in column 2.}
}
\value{
If the edges argument only includes one row, a vector of multiplet
   names is returned. If several edges are interogated a list is returned
   with one element per edge containing the names of the multiplets.
}
\description{
Returns the names of the multiplets that are associated with an edge.
}
\details{
Description
}
\examples{

e <- data.frame("A375", "HOS")
output <- getMultipletsForEdge(CIMseqSwarm_test, 1/10.5, e)

}
\author{
Jason T. Serviss
}
