% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\docType{methods}
\name{getMultipletsForEdge}
\alias{getMultipletsForEdge}
\alias{getMultipletsForEdge}
\alias{getMultipletsForEdge,spSwarm-method}
\title{getMultipletsForEdge}
\usage{
getMultipletsForEdge(spSwarm, ...)

\S4method{getMultipletsForEdge}{spSwarm}(spSwarm, edge.cutoff, edges, ...)
}
\arguments{
\item{spSwarm}{An spSwarm object.}

\item{...}{additional arguments to pass on}

\item{edge.cutoff}{The minimum fraction to consider (?).}

\item{edges}{A data frame indicating the edges of interest. Edges are
indicated by the nodes they connect with one node in column one and the
other node in column 2.}
}
\value{
If the edges argument only includes on row, a vector of multiplet
   names is returned. If several edges are interogated a list is returned
   with one element per edge containing the names of the multiplets.
}
\description{
Returns the names of the multiplets that are associated with an edge.
}
\details{
Description
}
\examples{

output <- getMultipletsForEdge(testSwa, 1/10.5, data.frame("A1", "B1"))

}
\author{
Jason T. Serviss
}
\keyword{getMultipletsForEdge}
