% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExtras.R
\name{longFormConnections}
\alias{longFormConnections}
\title{longFormConnections}
\usage{
longFormConnections(
  swarm,
  singlets,
  multiplets,
  maxCellsPerMultiplet = Inf,
  depleted = FALSE
)
}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}
}
\description{
Helper function for plotSwarmCircos.
}
\author{
Jason T. Serviss
}
