% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\name{pearsonsDist}
\alias{pearsonsDist}
\alias{pearsonsDist}
\title{pearsonsDist}
\usage{
pearsonsDist(spCounts, select)
}
\arguments{
\item{spCounts}{An spCounts object.}

\item{select}{A numeric vector indicating the indexes of genes to include.}
}
\value{
A matrix containing the mean value for each gene for each
   classification group.
}
\description{
Calculates the x and y coordinates of the mean of each classified group.
}
\details{
This method is typically only used in conjunction with plotting. It
calculates the 2 dimensional location of the mean of each classified group
in the supplied unsupervised dimensionality reduction (t-SNE) data
representation.
}
\examples{

s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])
my.dist <- pearsonsDist(cObjSng, 1:nrow(testCounts))

}
\author{
Jason T. Serviss
}
\keyword{pearsonsDist}
