% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\docType{methods}
\name{permuteSwarm}
\alias{permuteSwarm}
\alias{permuteSwarm}
\alias{permuteSwarm,spCounts-method}
\title{permuteSwarm}
\usage{
permuteSwarm(spCounts, ...)

\S4method{permuteSwarm}{spCounts}(spCounts, spCountsMul, spUnsupervised,
  spSwarm, distFun = distToSlice, maxiter = 10, swarmsize = 150,
  cores = 1, seed = 11, norm = TRUE, iter, edgeAlpha = 0.05, ...)
}
\arguments{
\item{...}{additional arguments to pass on}

\item{spSwarm}{An spSwarm object.}

\item{edge.cutoff}{The minimum fraction to consider (?).}

\item{multiplet}{The name of the multiplet of interest.}
}
\value{
Edge names.
}
\description{
Write something
}
\details{
Description
}
\examples{

#use demo data


#run function


}
\author{
Jason T. Serviss
}
\keyword{permuteSwarm}
