% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlot.R
\docType{methods}
\name{plotCounts}
\alias{plotCounts}
\alias{plotCounts}
\alias{plotCounts,spCounts-method}
\title{plotCounts}
\usage{
plotCounts(x, ...)

\S4method{plotCounts}{spCounts}(x, y, type, markers = NULL, ...)
}
\arguments{
\item{x}{An spCounts object containing singlets.}

\item{...}{additional arguments to pass on.}

\item{y}{An spCounts object containing multiplets.}

\item{type}{Can be "ercc", "markers", or .......}

\item{markers}{A character vector with 2 markers to plot.}
}
\value{
The spPlot function returns an object of class spCounts.
}
\description{
Subtitle
}
\details{
Imports count and count.ercc data to a sp.scRNAseq object.
}
\examples{

#use demo data
s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])
cObjMul <- spCounts(testCounts[, !s], testErcc[, !s])

#ERCC plot
p <- plotCounts(cObjSng, cObjMul, type = "ercc")

#Markers plot
markers <- c("a1", "a10")
p <- plotCounts(cObjSng, cObjMul, type = "markers", markers = markers)

}
\author{
Jason T. Serviss
}
\keyword{plotCounts}
