% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlot.R
\docType{methods}
\name{plotCountsData}
\alias{plotCountsData}
\alias{plotCountsData}
\alias{plotCountsData,spCounts-method}
\title{plotCountsData}
\usage{
plotCountsData(spCountsSng, ...)

\S4method{plotCountsData}{spCounts}(spCountsSng, spCountsMul, markers = NULL,
  ...)
}
\arguments{
\item{...}{additional arguments to pass on.}

\item{x}{An spUnsupervised object.}

\item{y}{An spCounts object containing singlets.}
}
\value{
A tibble with columns:
}
\description{
Assembles all data for plotCounts plots.
}
\examples{
#

}
\author{
Jason T. Serviss
}
\keyword{plotCountsData}
