% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts.R
\docType{methods}
\name{plotCountsData}
\alias{plotCountsData}
\alias{plotCountsData,CIMseqSinglets,CIMseqMultiplets-method}
\title{plotCountsData}
\usage{
plotCountsData(singlets, multiplets, ...)

\S4method{plotCountsData}{CIMseqSinglets,CIMseqMultiplets}(singlets,
  multiplets, markers = NULL, log = TRUE, normalize = TRUE, ...)
}
\arguments{
\item{singlets}{CIMseqSinglets; An CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; An CIMseqMultiplets object.}

\item{...}{additional arguments to pass on.}

\item{markers}{character; A vector with the 2 markers to plot. Must be
present in rownames of counts.}

\item{log}{logical; Use log2 + 1 values?}

\item{normalize}{logical; Use [0, 1] normalized values? 
See \code{\link{processMarkers}}.}
}
\value{
A tibble with columns:
}
\description{
Assembles all data for plotCounts plots. End users should use
the \code{\link{plotData}} function.
}
\examples{
d <- plotCountsData(CIMseqSinglets_test, CIMseqMultiplets_test)
}
\author{
Jason T. Serviss
}
\keyword{plotCountsData}
