% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotCounts.R
\name{plotCountsData}
\alias{plotCountsData}
\title{plotCountsData}
\usage{
plotCountsData(spCountsSng, spCountsMul, markers = NULL, ...)
}
\arguments{
\item{spCountsSng}{spCounts; An spCounts object containing singlets.}

\item{spCountsMul}{spCounts; An spCounts object containing multiplets.}

\item{markers}{character; A vector with the 2 markers to plot. Must be
present in rownames of counts.}

\item{...}{additional arguments to pass on.}
}
\value{
A tibble with columns:
}
\description{
Assembles all data for plotCounts plots. Not exported. End users should use
the \code{\link{plotData}} function.
}
\examples{

#use demo data
s <- grepl("^s", colnames(testCounts))
ercc <- grepl("^ERCC\\\\-[0-9]*$", rownames(testCounts))
cObjSng <- spCounts(testCounts[!ercc, s], testCounts[ercc, s])
cObjMul <- spCounts(testCounts[!ercc, !s], testCounts[ercc, !s])

d <- plotCountsData(cObjSng, cObjMul)
}
\author{
Jason T. Serviss
}
\keyword{plotCountsData}
