% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts.R
\docType{methods}
\name{plotCountsERCC}
\alias{plotCountsERCC}
\alias{plotCountsERCC,CIMseqSinglets,CIMseqMultiplets-method}
\title{plotCountsERCC}
\usage{
plotCountsERCC(singlets, multiplets, ...)

\S4method{plotCountsERCC}{CIMseqSinglets,CIMseqMultiplets}(singlets,
  multiplets, ...)
}
\arguments{
\item{singlets}{CIMseqSinglets; An CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; An CIMseqMultiplets object.}

\item{...}{additional arguments to pass on.}
}
\value{
A ggplot2 object containing the plot. See examples or the plotting
 vignette for further help.
}
\description{
Plot method for spCounts objects to display ERCC fraction and estimated cell
number per sample.
}
\examples{
p <- plotCountsERCC(CIMseqSinglets_test, CIMseqMultiplets_test)
}
\author{
Jason T. Serviss
}
