% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts.R
\docType{methods}
\name{plotCountsMarkers}
\alias{plotCountsMarkers}
\alias{plotCountsMarkers,CIMseqSinglets,CIMseqMultiplets-method}
\title{plotCountsMarkers}
\usage{
plotCountsMarkers(singlets, multiplets, ...)

\S4method{plotCountsMarkers}{CIMseqSinglets,CIMseqMultiplets}(singlets,
  multiplets, markers = NULL, log = TRUE, normalize = TRUE, ...)
}
\arguments{
\item{singlets}{CIMseqSinglets; An CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; An CIMseqMultiplets object.}

\item{...}{additional arguments to pass on.}

\item{markers}{character; A vector with the 2 markers to plot.}

\item{log}{logical; Use log2 values? See \code{\link{plotCountsData}}.}

\item{normalize}{logical; Use [0, 1] normalized values? 
See \code{\link{processMarkers}}.}
}
\value{
A ggplot2 object containing the plot. See examples or the plotting
 vignette for further help.
}
\description{
Plot method for spCounts objects to display "markers", typically genes
thought to be discreetly expressed in one cell type, in all samples.
}
\examples{
markers <- c("ACTB", "GAPDH")
p <- plotCountsMarkers(CIMseqSinglets_test, CIMseqMultiplets_test, markers)
}
\author{
Jason T. Serviss
}
