% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlot.R
\docType{methods}
\name{plotSwarm}
\alias{plotSwarm}
\alias{plotSwarm}
\alias{plotSwarm,spSwarm-method}
\title{plotSwarm}
\usage{
plotSwarm(x, ...)

\S4method{plotSwarm}{spSwarm}(x, y, z = NULL, w = NULL, type = NULL,
  loop = FALSE, markers = NULL, edge.cutoff = NULL, min.pval = NULL,
  min.num.edges = NULL, label.cutoff = NULL, ...)
}
\arguments{
\item{x}{An spSwarm object.}

\item{...}{additional arguments to pass on.}

\item{y}{An spUnsupervised object.}

\item{z}{An spCounts object with singlets.}

\item{w}{An spCounts object with multiplets.}

\item{type}{Specifies the plot type. Possible values are "tsne" or "igraph"
for network plots, "multiplets" or "edges" for residuals plots, "edgeBar"
or "pValueBar" for bar plots and "heat" for a heatmap.}

\item{loop}{Logical; TRUE if "self" connections should be plotted.}

\item{markers}{Not currently implemented.}

\item{edge.cutoff}{The minimum fraction to consider (?).}

\item{min.pval}{Minimum p-value to report.}

\item{min.num.edges}{Minimum number of observed edges to report a connection.}

\item{label.cutoff}{The number of labels to show in the residuals plot.}
}
\value{
The spPlot function returns an object of class spCounts.
}
\description{
Description.
}
\details{
Details.
}
\examples{

#use demo data
s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])
cObjMul <- spCounts(testCounts[, !s], testErcc[, !s])
uObj <- testUns
sObj <- testSwa

# tsne plot
p <- plotSwarm(sObj, uObj, cObjSng, cObjMul, type = "tsne")

#igraph plot
p <- plotSwarm(sObj, uObj, cObjSng, cObjMul, type = "igraph")

#multiplet residuals plot
p <- plotSwarm(sObj, uObj, cObjSng, cObjMul, type = "multiplets")

#edge residuals plot
p <- plotSwarm(sObj, uObj, cObjSng, cObjMul, type = "edges")

#edge barplot
p <- plotSwarm(sObj, uObj, cObjSng, cObjMul, type = "edgeBar")

#pvalue barplot
p <- plotSwarm(sObj, uObj, cObjSng, cObjMul, type = "pValueBar")

#heatmap plot
p <- plotSwarm(sObj, uObj, cObjSng, cObjMul, type = "heat")

}
\author{
Jason T. Serviss
}
\keyword{plotSwarm}
