% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSwarm.R
\docType{methods}
\name{plotSwarmBarBase}
\alias{plotSwarmBarBase}
\alias{plotSwarmBarBase,CIMseqSwarm,CIMseqSinglets-method}
\title{plotSwarmBarBase}
\usage{
plotSwarmBarBase(swarm, singlets, ...)

\S4method{plotSwarmBarBase}{CIMseqSwarm,CIMseqSinglets}(swarm, singlets,
  ...)
}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{...}{additional arguments to pass on.}
}
\value{
A ggplot object with the base data.
}
\description{
Not exported. Provides the base plot for all spSwarm bar type plots.
}
\author{
Jason T. Serviss
}
