% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotSwarm.R
\docType{methods}
\name{plotSwarmBarBase}
\alias{plotSwarmBarBase}
\alias{plotSwarmBarBase,spSwarm-method}
\title{plotSwarmBarBase}
\usage{
plotSwarmBarBase(spSwarm, ...)

\S4method{plotSwarmBarBase}{spSwarm}(spSwarm, ...)
}
\arguments{
\item{spSwarm}{spSwarm; An spSwarm object.}

\item{...}{additional arguments to pass on.}
}
\value{
A ggplot object with the base data.
}
\description{
Not exported. Provides the base plot for all spSwarm bar type plots.
}
\author{
Jason T. Serviss
}
\keyword{plotSwarmBarBase}
