% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSwarm.R
\docType{methods}
\name{plotSwarmCircos}
\alias{plotSwarmCircos}
\alias{plotSwarmCircos,CIMseqSwarm-method}
\title{plotSwarmGenes}
\usage{
plotSwarmCircos(swarm, ...)

\S4method{plotSwarmCircos}{CIMseqSwarm}(swarm, singlets, multiplets,
  classOrder = NULL, connectionClass = NULL, alpha = 0.05,
  weightCut = 10, label.cex = 1, legend = TRUE,
  pal = colorRampPalette(c("grey90", viridis::viridis(1)))(120)[20:110],
  nonSigCol = "grey90", ...)
}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{...}{additional arguments to pass on.}

\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}

\item{classOrder}{character; Order of the cell types / classes on the circos.}

\item{connectionClass}{character; Vector of length 1 specifying the class /
cell type to plot or NULL to plot all classes / cell types.}

\item{alpha}{numeric; Vector of length 1 specifying the which p-values are 
considered significant.}

\item{weightCut}{integer; Vector of length 1 specifying weights below which 
the p-value should not be calculated.}

\item{label.cex}{numeric; Vector of length 1 between [0, 1] indicating the
size of the cell type labels.}

\item{legend}{logical; indicates if the legends should be plotted.}

\item{pal}{character; A vector including the colour pallete for the score 
colours.}

\item{nonSigCol}{character; Vector of length 1 indicating the colours for 
non-significant connections.}
}
\value{
A ggplot object.
}
\description{
plotSwarmGenes
}
\author{
Jason T. Serviss
}
