% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotSwarm.R
\docType{methods}
\name{plotSwarmEdgeBar}
\alias{plotSwarmEdgeBar}
\alias{plotSwarmEdgeBar,spSwarm-method}
\title{plotSwarmEdgeBar}
\usage{
plotSwarmEdgeBar(spSwarm, ...)

\S4method{plotSwarmEdgeBar}{spSwarm}(spSwarm, ...)
}
\arguments{
\item{spSwarm}{spSwarm; An spSwarm object.}

\item{...}{additional arguments to pass on.}
}
\value{
A ggplot object.
}
\description{
plotSwarmEdgeBar
}
\author{
Jason T. Serviss
}
\keyword{plotSwarmEdgeBar}
