% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSwarm.R
\docType{methods}
\name{plotSwarmEdgeBar}
\alias{plotSwarmEdgeBar}
\alias{plotSwarmEdgeBar,CIMseqSwarm-method}
\title{plotSwarmEdgeBar}
\usage{
plotSwarmEdgeBar(swarm, ...)

\S4method{plotSwarmEdgeBar}{CIMseqSwarm}(swarm, ...)
}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{...}{additional arguments to pass on.}
}
\value{
A ggplot object.
}
\description{
plotSwarmEdgeBar
}
\author{
Jason T. Serviss
}
