% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSwarm.R
\docType{methods}
\name{plotSwarmGenes}
\alias{plotSwarmGenes}
\alias{plotSwarmGenes,CIMseqSwarm-method}
\title{plotSwarmGenes}
\usage{
plotSwarmGenes(swarm, ...)

\S4method{plotSwarmGenes}{CIMseqSwarm}(swarm, singlets, multiplets, genes,
  multipletsToPlot, freq = 10, ...)
}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{...}{additional arguments to pass on.}

\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}

\item{genes}{Character; Genes to be plotted. Can not exceed 20.}

\item{multipletsToPlot}{Character; Multiplets to be plotted.}

\item{freq}{Numeric, Length 1 vector indicating the frequency the cost should
be calculated along x.}
}
\value{
A ggplot object.
}
\description{
plotSwarmGenes
}
\author{
Jason T. Serviss
}
\keyword{plotSwarmGenes}
