% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotSwarm.R
\docType{methods}
\name{plotSwarmGenes}
\alias{plotSwarmGenes}
\alias{plotSwarmGenes}
\alias{plotSwarmGenes,spSwarm-method}
\title{plotSwarmGenes}
\usage{
plotSwarmGenes(spSwarm, ...)

\S4method{plotSwarmGenes}{spSwarm}(spSwarm, spCountsMul, genes, multiplets,
  freq = 10, ...)
}
\arguments{
\item{spSwarm}{spSwarm; An spSwarm object.}

\item{...}{additional arguments to pass on.}

\item{genes}{Character; Genes to be plotted. Can not exceed 20.}

\item{multiplets}{Character; Multiplets to be plotted.}

\item{freq}{Numeric, Length 1 vector indicating the frequency the cost should
be calculated along x.}

\item{spCountsMul;}{An spCounts object containing multiplets.}

\item{fractions}{Can either be numeric vector, in which case the fractions
are applied to all of the samples, a data.frame mimicking the structure of
the spSwarm slot in an spSwarm object, or NULL, in which case the fractions
are extracted from the spSwarm argument.}
}
\value{
A ggplot object.
}
\description{
plotSwarmGenes
}
\author{
Jason T. Serviss
}
\keyword{plotSwarmGenes}
