% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSwarm.R
\docType{methods}
\name{plotSwarmGraph}
\alias{plotSwarmGraph}
\alias{plotSwarmGraph,CIMseqSwarm,CIMseqSinglets-method}
\title{plotSwarmGraph}
\usage{
plotSwarmGraph(swarm, singlets, ...)

\S4method{plotSwarmGraph}{CIMseqSwarm,CIMseqSinglets}(swarm, singlets, ...)
}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{...}{additional arguments to pass on.}
}
\value{
The spPlot function returns an object of class spCounts.
}
\description{
Description.
}
\details{
Details.
}
\examples{

p <- plotSwarmGraph(CIMseqSwarm_test, CIMseqSinglets_test)

}
\author{
Jason T. Serviss
}
\keyword{plotSwarmGraph}
