% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotSwarm.R
\docType{methods}
\name{plotSwarmGraph}
\alias{plotSwarmGraph}
\alias{plotSwarmGraph}
\alias{plotSwarmGraph,spSwarm-method}
\title{plotSwarmGraph}
\usage{
plotSwarmGraph(spSwarm, ...)

\S4method{plotSwarmGraph}{spSwarm}(spSwarm, ...)
}
\arguments{
\item{spSwarm}{spSwarm; An spSwarm object.}

\item{...}{additional arguments to pass on.}
}
\value{
The spPlot function returns an object of class spCounts.
}
\description{
Description.
}
\details{
Details.
}
\examples{

#use demo data
s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])
cObjMul <- spCounts(testCounts[, !s], testErcc[, !s])
uObj <- testUns
sObj <- testSwa

#plot
p <- plotSwarmGraph(sObj)
}
\author{
Jason T. Serviss
}
\keyword{plotSwarmGraph}
