% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotSwarm.R
\docType{methods}
\name{plotSwarmGraph}
\alias{plotSwarmGraph}
\alias{plotSwarmGraph,spSwarm,spUnsupervised-method}
\title{plotSwarmGraph}
\usage{
plotSwarmGraph(spSwarm, spUnsupervised, ...)

\S4method{plotSwarmGraph}{spSwarm,spUnsupervised}(spSwarm, spUnsupervised,
  ...)
}
\arguments{
\item{spSwarm}{spSwarm; An spSwarm object.}

\item{spUnsupervised}{spUnsupervised; An spUnsupervised object.}

\item{...}{additional arguments to pass on.}
}
\value{
The spPlot function returns an object of class spCounts.
}
\description{
Description.
}
\details{
Details.
}
\examples{

p <- plotSwarmGraph(test_spSwarm, test_spUnsupervised)

}
\author{
Jason T. Serviss
}
\keyword{plotSwarmGraph}
