% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotSwarm.R
\docType{methods}
\name{plotSwarmHeat}
\alias{plotSwarmHeat}
\alias{plotSwarmHeat,spSwarm-method}
\title{plotSwarmHeat}
\usage{
plotSwarmHeat(spSwarm, ...)

\S4method{plotSwarmHeat}{spSwarm}(spSwarm, ...)
}
\arguments{
\item{spSwarm}{spSwarm; An spSwarm object.}

\item{...}{additional arguments to pass on.}
}
\value{
A ggplot object.
}
\description{
plotSwarmHeat
}
\author{
Jason T. Serviss
}
\keyword{plotSwarmHeat}
