% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSwarm.R
\name{plotSwarmHeat}
\alias{plotSwarmHeat}
\alias{plotSwarmHeat,CIMseqSwarm,CIMseqSinglets,CIMseqMultiplets-method}
\title{plotSwarmHeat}
\usage{
plotSwarmHeat(swarm, singlets, multiplets, ...)

\S4method{plotSwarmHeat}{CIMseqSwarm,CIMseqSinglets,CIMseqMultiplets}(swarm, singlets, multiplets, ...)
}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}

\item{...}{additional arguments to pass on.}
}
\value{
A ggplot object.
}
\description{
plotSwarmHeat
}
\author{
Jason T. Serviss
}
\keyword{plotSwarmHeat}
