% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSwarm.R
\docType{methods}
\name{plotSwarmHeat}
\alias{plotSwarmHeat}
\alias{plotSwarmHeat,CIMseqSwarm-method}
\title{plotSwarmHeat}
\usage{
plotSwarmHeat(swarm, ...)

\S4method{plotSwarmHeat}{CIMseqSwarm}(swarm, ...)
}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{...}{additional arguments to pass on.}
}
\value{
A ggplot object.
}
\description{
plotSwarmHeat
}
\author{
Jason T. Serviss
}
\keyword{plotSwarmHeat}
