% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotSwarm.R
\docType{methods}
\name{plotSwarmPbar}
\alias{plotSwarmPbar}
\alias{plotSwarmPbar}
\alias{plotSwarmPbar,spSwarm-method}
\title{plotSwarmPbar}
\usage{
plotSwarmPbar(spSwarm, ...)

\S4method{plotSwarmPbar}{spSwarm}(spSwarm, ...)
}
\arguments{
\item{spSwarm}{spSwarm; An spSwarm object.}

\item{...}{additional arguments to pass on.}
}
\value{
A ggplot object.
}
\description{
plotSwarmPbar
}
\author{
Jason T. Serviss
}
\keyword{plotSwarmPbar}
