% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSwarm.R
\name{plotSwarmPbar}
\alias{plotSwarmPbar}
\alias{plotSwarmPbar,CIMseqSwarm,CIMseqSinglets,CIMseqMultiplets-method}
\title{plotSwarmPbar}
\usage{
plotSwarmPbar(swarm, singlets, multiplets, ...)

\S4method{plotSwarmPbar}{CIMseqSwarm,CIMseqSinglets,CIMseqMultiplets}(swarm, singlets, multiplets, ...)
}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}

\item{...}{additional arguments to pass on.}
}
\value{
A ggplot object.
}
\description{
plotSwarmPbar
}
\author{
Jason T. Serviss
}
\keyword{plotSwarmPbar}
