% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging.R
\docType{methods}
\name{plotSwarmPosition}
\alias{plotSwarmPosition}
\alias{plotSwarmPosition}
\alias{plotSwarmPosition,spSwarm-method}
\title{plotSwarmPosition}
\usage{
plotSwarmPosition(x, ...)

\S4method{plotSwarmPosition}{spSwarm}(x, multiplet, interval = NULL, ...)
}
\arguments{
\item{x}{An spSwarm object.}

\item{...}{additional arguments to pass on.}

\item{multiplet}{The name of the multiplet to plot the swarm positions for.}

\item{interval}{The iterations to be plotted if all recorded iteration
intervals should not be plotted.}
}
\value{
The plotSwarmPosition function returns an object of class spCounts.
}
\description{
Subtitle
}
\details{
Plots the swarm position in a spSwarm object for each reported iteration
(specified by reportRate) when report equals TRUE.
}
\examples{

#no test data where swarm was tracked with report arg.
}
\author{
Jason T. Serviss
}
\keyword{plotSwarmPosition}
