% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlot.R
\docType{methods}
\name{plotUnsupervised}
\alias{plotUnsupervised}
\alias{plotUnsupervised}
\alias{plotUnsupervised,spUnsupervised-method}
\title{plotUnsupervised}
\usage{
plotUnsupervised(x, ...)

\S4method{plotUnsupervised}{spUnsupervised}(x, y = NULL, type = "clusters",
  markers = NULL, plotUncertainty = TRUE, ...)
}
\arguments{
\item{x}{An spUnsupervised object.}

\item{...}{additional arguments to pass on.}

\item{y}{An spCounts object containing singlets.}

\item{type}{Can be "clusters", "markers", or .......}

\item{markers}{A character vector with markers to plot.}

\item{plotUncertainty}{Logical indicating if uncertainty should be
represented as the point size in the plots.}
}
\value{
The plotUnsupervised function returns an object of class spCounts.
}
\description{
Subtitle
}
\details{
Imports count and count.ercc data to a sp.scRNAseq object.
}
\examples{

#use demo data
s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])
uObj <- testUns

#Clusters plot
p <- plotUnsupervised(uObj, cObjSng, type = "clusters")

#Markers plot
markers <- c("a1", "a10")
p <- plotUnsupervised(uObj, cObjSng, type = "markers", markers = markers)

}
\author{
Jason T. Serviss
}
\keyword{plotUnsupervised}
