% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotUnsupervised.R
\docType{methods}
\name{plotUnsupervisedClass}
\alias{plotUnsupervisedClass}
\alias{plotUnsupervisedClass,spUnsupervised,spCounts-method}
\title{plotUnsupervisedClass}
\usage{
plotUnsupervisedClass(spUnsupervised, spCountsSng, ...)

\S4method{plotUnsupervisedClass}{spUnsupervised,spCounts}(spUnsupervised,
  spCountsSng, ...)
}
\arguments{
\item{spUnsupervised}{spUnsupervised; An spUnsupervised object.}

\item{spCountsSng}{spCounts; An spCounts object containing singlets.}

\item{...}{additional arguments to pass on.}
}
\value{
The ggplot2 object containing the plot. See examples or the plotting
 vignette for further help.
}
\description{
Plot method for spUnsupervised objects to highlight dimensionality reduction
and classification results.
}
\examples{

#use demo data
s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])
uObj <- testUns

#Plot
p <- plotUnsupervisedClass(uObj, cObjSng)
}
\author{
Jason T. Serviss
}
\keyword{plotUnsupervisedClass}
