% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotData.R
\docType{methods}
\name{plotUnsupervisedData}
\alias{plotUnsupervisedData}
\alias{plotUnsupervisedData}
\alias{plotUnsupervisedData,spUnsupervised-method}
\title{plotUnsupervisedData}
\usage{
plotUnsupervisedData(spUnsupervised, ...)

\S4method{plotUnsupervisedData}{spUnsupervised}(spUnsupervised, spCountsSng,
  markers = NULL, pal = NULL, ...)
}
\arguments{
\item{spUnsupervised}{spUnsupervised; An spUnsupervised object.}

\item{...}{additional arguments to pass on.}

\item{spCountsSng}{spCounts; An spCounts object containing singlets.}
}
\value{
A tibble with columns:
}
\description{
Assembles all data for plotUnsupervised plots.
}
\examples{

#use demo data
s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])
uObj <- testUns

plotUnsupervisedData(sObjSng, uObj)
}
\author{
Jason T. Serviss
}
\keyword{plotUnsupervisedData}
