% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlot.R
\docType{methods}
\name{plotUnsupervisedData}
\alias{plotUnsupervisedData}
\alias{plotUnsupervisedData}
\alias{plotUnsupervisedData,spUnsupervised-method}
\title{plotUnsupervisedData}
\usage{
plotUnsupervisedData(x, ...)

\S4method{plotUnsupervisedData}{spUnsupervised}(x, y, markers = NULL,
  pal = NULL, ...)
}
\arguments{
\item{x}{An spUnsupervised object.}

\item{...}{additional arguments to pass on.}

\item{y}{An spCounts object containing singlets.}
}
\value{
A tibble with columns:
}
\description{
Assembles all data for plotUnsupervised plots.
}
\examples{
#

}
\author{
Jason T. Serviss
}
\keyword{plotUnsupervisedData}
