% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts.R
\docType{methods}
\name{plotUnsupervisedMarkers}
\alias{plotUnsupervisedMarkers}
\alias{plotUnsupervisedMarkers,CIMseqSinglets,CIMseqMultiplets-method}
\title{plotUnsupervisedMarkers}
\usage{
plotUnsupervisedMarkers(singlets, multiplets, ...)


  \S4method{plotUnsupervisedMarkers}{CIMseqSinglets,CIMseqMultiplets}(singlets,
  multiplets, markers = NULL, log = TRUE, normalize = TRUE,
  pal = col40(), ...)
}
\arguments{
\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{multiplets}{CIMseqMultiplets; A CIMseqMultiplets object.}

\item{...}{additional arguments to pass on.}

\item{markers}{character; A vector with markers to be included plot.}

\item{log}{logical; Use log2 values? See \code{\link{plotCountsData}}.}

\item{normalize}{logical; Use [0, 1] normalized values? 
See \code{\link{processMarkers}}.}

\item{pal}{character; A palette of colors with length(pal) = length(markers).}
}
\value{
A ggplot2 object with plot. See examples or the plotting vignette
for further help.
}
\description{
Plot method for spUnsupervised objects to highlight "markers", typically
genes thought to be discreetly expressed in one cell type, in the context of
the dimensionality reduction and classification results.
}
\examples{

#Plot
p <- plotUnsupervisedMarkers(
  CIMseqSinglets_test, CIMseqMultiplets_test, markers = "CD74"
)
}
\author{
Jason T. Serviss
}
