% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlotUnsupervised.R
\docType{methods}
\name{plotUnsupervisedMarkers}
\alias{plotUnsupervisedMarkers}
\alias{plotUnsupervisedMarkers,spUnsupervised,spCounts-method}
\title{plotUnsupervisedMarkers}
\usage{
plotUnsupervisedMarkers(spUnsupervised, spCountsSng, ...)


  \S4method{plotUnsupervisedMarkers}{spUnsupervised,spCounts}(spUnsupervised,
  spCountsSng, markers = NULL, pal = col40(), ...)
}
\arguments{
\item{spUnsupervised}{spUnsupervised; An spUnsupervised object.}

\item{spCountsSng}{spCounts; An spCounts object containing singlets.}

\item{...}{additional arguments to pass on.}

\item{markers}{character; A vector with markers to be included plot.}

\item{pal}{character; A palette of colors with length(pal) = length(markers).}
}
\value{
A ggplot2 object with plot. See examples or the plotting vignette
for further help.
}
\description{
Plot method for spUnsupervised objects to highlight "markers", typically
genes thought to be discreetly expressed in one cell type, in the context of
the dimensionality reduction and classification results.
}
\examples{

#use demo data
s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])
uObj <- testUns

#Plot
p <- plotUnsupervisedMarkers(uObj, cObjSng, markers = "a10")
}
\author{
Jason T. Serviss
}
\keyword{plotUnsupervisedMarkers}
