% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\name{runMclust}
\alias{runMclust}
\alias{runMclust}
\title{runMclust}
\usage{
runMclust(my.tsne, Gmax = 50, seed = 11)
}
\arguments{
\item{my.tsne}{A tsne result.}

\item{Gmax}{A integer indicating the maximum number of clusters to evaluate.}

\item{seed}{The desired seed to set before running.}
}
\value{
A matrix containing the mean value for each gene for each
   classification group.
}
\description{
Calculates the x and y coordinates of the mean of each classified group.
}
\details{
This method is typically only used in conjunction with plotting. It
calculates the 2 dimensional location of the mean of each classified group
in the supplied unsupervised dimensionality reduction (t-SNE) data
representation.
}
\examples{

s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])
select <- spTopMax(cObjSng, 10)
my.dist <- pearsonsDist(cObjSng, select)
tsne <- runTsne(my.dist, max_iter = 10)
class <- runMclust(tsne, 6, 11)

}
\author{
Jason T. Serviss
}
\keyword{runMclust}
