% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{runPCA}
\alias{runPCA}
\title{runPCA}
\usage{
runPCA(singlets, select)
}
\arguments{
\item{singlets}{CIMseqSinglets; A CIMseqSinglets object.}

\item{select}{A numeric vector indicating the indexes of genes to include.}
}
\description{
Runs a principal component analysis using \code{\link[gmodels]{fast.prcomp}}.
}
\author{
Jason T. Serviss
}
