% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-classes.R, R/initialize-methods.R, R/process-input.R, R/show-methods.R
\docType{class}
\name{spCounts-class}
\alias{.Counts}
\alias{getData}
\alias{getData,spCounts-method}
\alias{initialize,spCounts-method}
\alias{show,spCounts-method}
\alias{spCounts}
\alias{spCounts,matrix-method}
\alias{spCounts-class}
\title{spCounts}
\usage{
getData(x, ...)

\S4method{getData}{spCounts}(x, n = NULL)

getData(x, ...)

\S4method{getData}{spCounts}(x, n = NULL)

\S4method{initialize}{spCounts}(.Object, ..., counts, counts.log, counts.ercc,
  sampleType)

spCounts(counts, ...)

\S4method{spCounts}{matrix}(counts, counts.ercc, sampleType = "1000102901",
  ...)

\S4method{show}{spCounts}(object)
}
\arguments{
\item{...}{additional arguments to pass on}

\item{counts}{Counts matrix with samples as columns and genes as rows.}
}
\value{
The spCounts function returns an object of class spCounts.
}
\description{
Subtitle
}
\details{
Imports count and count.ercc data to a sp.scRNAseq object.
}
\examples{

#use demo data
data(Doublet_project_data)

#run function


}
\author{
Jason T. Serviss
}
\keyword{spCounts}

