% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-classes.R, R/initialize-methods.R, R/show-methods.R, R/spCounts.R
\docType{class}
\name{spSwarm-class}
\alias{.Counts}
\alias{getData}
\alias{getData,spCounts-method}
\alias{initialize,spCounts-method}
\alias{show,spCounts-method}
\alias{spCounts}
\alias{spCounts,matrix-method}
\alias{spSwarm-class}
\title{spCounts}
\usage{
getData(x, ...)

\S4method{getData}{spCounts}(x, n = NULL)

\S4method{initialize}{spCounts}(.Object, ..., counts, counts.log, counts.ercc,
  sampleType)

\S4method{show}{spCounts}(object)

spCounts(counts, ...)

\S4method{spCounts}{matrix}(counts, counts.ercc, sampleType = "1000102901",
  ...)
}
\arguments{
\item{...}{additional arguments to pass on}

\item{counts}{Counts matrix with samples as columns and genes as rows.}

\item{counts.ercc}{A matrix containing ercc spike-in reads and their counts.}

\item{sampleType}{A character vector indicating the column naming scheme showing that the column contains a multuplet.}
}
\value{
The spCounts function returns an object of class spCounts.
}
\description{
Subtitle
}
\details{
Imports count, sampleType, and count.ercc data to a sp.scRNAseq object.
}
\examples{

#use demo data
data(expData)
counts <- getData(expData, "counts")
counts.ercc <- getData(expData, "counts.ercc")
sampleType <- '1000102901' #indicates that the column contains a multuplet

#run function
cObj <- spCounts(counts, counts.ercc, '1000102901')

}
\author{
Jason T. Serviss
}
\keyword{spCounts}

