% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-classes.R, R/initialize-methods.R,
%   R/show-methods.R, R/spCounts.R
\docType{class}
\name{spCounts-class}
\alias{spCounts-class}
\alias{.spCounts}
\alias{getData}
\alias{getData,spCounts-method}
\alias{initialize,spCounts-method}
\alias{show,spCounts-method}
\alias{spCounts}
\alias{spCounts}
\alias{spCounts,matrix-method}
\title{spCounts}
\usage{
getData(object, ...)

\S4method{getData}{spCounts}(object, n = NULL)

\S4method{initialize}{spCounts}(.Object, ..., counts, counts.log, counts.cpm,
  counts.ercc)

\S4method{show}{spCounts}(object)

spCounts(counts, ...)

\S4method{spCounts}{matrix}(counts, counts.ercc, ...)
}
\arguments{
\item{object}{spCounts object.}

\item{...}{additional arguments to pass on}

\item{n}{Data to extract from spCounts object.}

\item{.Object}{Internal object.}

\item{counts}{Counts matrix with samples as columns and genes as rows.}

\item{counts.log}{Log2 normalized counts per million.}

\item{counts.cpm}{Normalized counts per million.}

\item{counts.ercc}{A matrix containing ercc spike-in reads and their counts.}
}
\value{
spCounts object.
}
\description{
Subtitle
}
\details{
Imports count, sampleType, and count.ercc data to a sp.scRNAseq object.
}
\examples{

s <- grepl("^s", colnames(testCounts))
cObj <- spCounts(testCounts[, s], testErcc[, s])

}
\author{
Jason T. Serviss
}
\keyword{spCounts}
