% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{spPlot}
\alias{spPlot}
\alias{spPlot,spCounts-method}
\title{spPlot}
\usage{
spPlot(counts, ...)

\S4method{spPlot}{spCounts}(counts, type, markers = NULL, ...)
}
\arguments{
\item{counts}{Counts matrix with samples as columns and genes as rows.}

\item{...}{additional arguments to pass on}

\item{type}{Can be "ercc", "markers", or .......}

\item{markers}{A character vector with 2 markers to plot.}
}
\value{
The spPlot function returns an object of class spCounts.
}
\description{
Subtitle
}
\details{
Imports count and count.ercc data to a sp.scRNAseq object.
}
\examples{

#use demo data
data(Doublet_project_data)

#run function
spPlot(x, type="ercc")

}
\author{
Jason T. Serviss
}
\keyword{spPlot}

