% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPlot.R
\docType{methods}
\name{spPlot}
\alias{spPlot}
\alias{spPlot,spCounts-method}
\alias{spPlot,spSwarm-method}
\alias{spPlot,spUnsupervised-method}
\title{spPlot}
\usage{
spPlot(x, ...)

\S4method{spPlot}{spCounts}(x, type, markers = NULL, ...)

\S4method{spPlot}{spUnsupervised}(x, type, markers = NULL, ...)

\S4method{spPlot}{spSwarm}(x, layout = "tsne", loop = TRUE, ...)
}
\arguments{
\item{x}{Counts matrix with samples as columns and genes as rows.}

\item{...}{additional arguments to pass on}

\item{type}{Can be "ercc", "markers", or .......}

\item{markers}{A character vector with 2 markers to plot.}

\item{layout}{Specify "tsne" for overlaying connections over the tSNE or igraph for graph layout.}

\item{loop}{Logical; TRUE if "self" connections should be turned on.}
}
\value{
The spPlot function returns an object of class spCounts.
}
\description{
Subtitle
}
\details{
Imports count and count.ercc data to a sp.scRNAseq object.
}
\examples{

#use demo data
data(expData)

#run function

}
\author{
Jason T. Serviss
}
\keyword{spPlot}

