% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-classes.R, R/initialize-methods.R,
%   R/show-methods.R, R/spSwarm.R
\docType{class}
\name{spSwarm-class}
\alias{spSwarm-class}
\alias{.spSwarm}
\alias{getData,spSwarm-method}
\alias{initialize,spSwarm-method}
\alias{show,spSwarm-method}
\alias{spSwarm}
\alias{spSwarm}
\alias{spSwarm,spCounts,spUnsupervised-method}
\title{spSwarm}
\usage{
\S4method{getData}{spSwarm}(object, n = NULL)

\S4method{initialize}{spSwarm}(.Object, ..., spSwarm, costs, convergence,
  arguments)

\S4method{show}{spSwarm}(object)

spSwarm(spCounts, spUnsupervised, ...)

\S4method{spSwarm}{spCounts,spUnsupervised}(spCounts, spUnsupervised,
  distFun = distToSlice, maxiter = 10, swarmsize = 150, cores = 1,
  seed = 11, ...)
}
\arguments{
\item{object}{spRSwarm object.}

\item{n}{Data to extract from spRSwarm object.}

\item{.Object}{Internal object.}

\item{...}{additional arguments to pass on}

\item{spSwarm}{The spSwarm results.}

\item{costs}{The costs after optimization.}

\item{convergence}{The convergence output from psoptim. One value per multiplet.}

\item{arguments}{Arguments passed to the spRSwarm function.}

\item{spCounts}{an spCount object with multiplets.}

\item{spUnsupervised}{an spCount object.}

\item{distFun}{The distance function used to calculate the cost.}

\item{maxiter}{pySwarm argument indicating maximum optimization iterations.}

\item{swarmsize}{pySwarm argument indicating the number of swarm particals.}

\item{cores}{The number of cores to be used while running spRSwarm.}
}
\value{
spSwarm output.
}
\description{
Subtitle
}
\details{
Description
}
\examples{

#use demo data


#run function


}
\author{
Jason T. Serviss
}
\keyword{spSwarm}
