% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-classes.R, R/initialize-methods.R,
%   R/show-methods.R, R/spSwarm.R
\docType{class}
\name{spSwarm-class}
\alias{spSwarm-class}
\alias{.spSwarm}
\alias{getData,spSwarm-method}
\alias{initialize,spSwarm-method}
\alias{show,spSwarm-method}
\alias{spSwarm}
\alias{spSwarm}
\alias{spSwarm,spCounts,spCounts,spUnsupervised-method}
\title{spSwarm}
\usage{
\S4method{getData}{spSwarm}(object, n = NULL)

\S4method{initialize}{spSwarm}(.Object, ..., spSwarm, costs, convergence, stats,
  arguments)

\S4method{show}{spSwarm}(object)

spSwarm(spCountsSng, spCountsMul, spUnsupervised, ...)

\S4method{spSwarm}{spCounts,spCounts,spUnsupervised}(spCountsSng, spCountsMul,
  spUnsupervised, maxiter = 10, swarmsize = 150,
  nSyntheticMultiplets = 200, cores = 1, seed = 11, norm = TRUE,
  report = FALSE, reportRate = NULL, selectInd = NULL,
  vectorize = FALSE, ...)
}
\arguments{
\item{object}{spRSwarm object.}

\item{n}{Data to extract from spRSwarm object.}

\item{.Object}{Internal object.}

\item{...}{additional arguments to pass on}

\item{spSwarm}{The spSwarm results.}

\item{costs}{The costs after optimization.}

\item{convergence}{The convergence output from psoptim. One value per
multiplet.}

\item{stats}{The stats output from psoptim.}

\item{arguments}{Arguments passed to the spSwarm function.}

\item{spCountsSng}{an spCount object with singlets.}

\item{spCountsMul}{an spCount object with multiplets.}

\item{spUnsupervised}{an spCount object.}

\item{maxiter}{pySwarm argument indicating maximum optimization iterations.}

\item{swarmsize}{pySwarm argument indicating the number of swarm particals.}

\item{nSyntheticMultiplets}{Numeric value indicating the number of synthetic
multiplets to generate during deconvolution.}

\item{cores}{The number of cores to be used while running spRSwarm.}

\item{seed}{The desired seed to set before running.}

\item{norm}{Logical indicating if the sum of fractions should equal 1.}

\item{report}{Logical indicating if additional reporting from the
optimization should be included.}

\item{reportRate}{If report is TRUE, the iteration interval that a report
should be generated.}

\item{selectInd}{Numeric; Gene indexes to select for swarm optimization. If
NULL the selectInd slot from the spUnsupervised object is used.}

\item{vectorize}{Argument to \link[pso]{psoptim}.}
}
\value{
spSwarm output.
}
\description{
Subtitle
}
\details{
Description
}
\examples{

#use demo data

}
\author{
Jason T. Serviss
}
\keyword{spSwarm}
