% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-classes.R, R/initialize-methods.R,
%   R/show-methods.R, R/spSwarm.R
\docType{class}
\name{spSwarm-class}
\alias{spSwarm-class}
\alias{.spSwarm}
\alias{getData,spSwarm-method}
\alias{initialize,spSwarm-method}
\alias{show,spSwarm-method}
\alias{spSwarm}
\alias{spSwarm}
\alias{spSwarm,spCounts,spUnsupervised-method}
\title{spSwarm}
\usage{
\S4method{getData}{spSwarm}(object, n = NULL)

\S4method{initialize}{spSwarm}(.Object, ..., spSwarm, costs, convergence, stats,
  arguments)

\S4method{show}{spSwarm}(object)

spSwarm(spCounts, spUnsupervised, ...)

\S4method{spSwarm}{spCounts,spUnsupervised}(spCounts, spUnsupervised,
  distFun = "distToSlice", maxiter = 10, swarmsize = 150, cores = 1,
  seed = 11, norm = TRUE, report = FALSE, reportRate = NULL,
  cellNumbers = NULL, e = NULL, ...)
}
\arguments{
\item{object}{spRSwarm object.}

\item{n}{Data to extract from spRSwarm object.}

\item{.Object}{Internal object.}

\item{...}{additional arguments to pass on}

\item{spSwarm}{The spSwarm results.}

\item{costs}{The costs after optimization.}

\item{convergence}{The convergence output from psoptim. One value per
multiplet.}

\item{stats}{The stats output from psoptim.}

\item{arguments}{Arguments passed to the spSwarm function.}

\item{spCounts}{an spCount object with multiplets.}

\item{spUnsupervised}{an spCount object.}

\item{distFun}{The distance function used to calculate the cost. Either the
name of a custom function in the local environment or one of the included
functions, i.e. \code{distToSlice, distToSliceNorm, distToSliceTop,
distToSliceEuclid, distToSlicePearson, bic}.}

\item{maxiter}{pySwarm argument indicating maximum optimization iterations.}

\item{swarmsize}{pySwarm argument indicating the number of swarm particals.}

\item{cores}{The number of cores to be used while running spRSwarm.}

\item{seed}{The desired seed to set before running.}

\item{norm}{Logical indicating if the sum of fractions should equal 1.}

\item{report}{Logical indicating if additional reporting from the
optimization should be included.}

\item{reportRate}{If report is TRUE, the iteration interval that a report
should be generated.}

\item{cellNumbers}{Tibble; Output from \code{estimateCells} function.}

\item{e}{Numeric; The epsilon value for the .complexityPenilty unit.}
}
\value{
spSwarm output.
}
\description{
Subtitle
}
\details{
Description
}
\examples{

#use demo data
s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])
cObjMul <- spCounts(testCounts[, !s], testErcc[, !s])
cn <- estimateCells(cObjSng, cObjMul)
sObj <- spSwarm(cObjMul, testUns, distFun = "dtsnCellNum", cellNumbers = cn, e = 0.0025)

}
\author{
Jason T. Serviss
}
\keyword{spSwarm}
