% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spSwarm.R
\name{spSwarmPoisson}
\alias{spSwarmPoisson}
\alias{spSwarmPoisson}
\title{spSwarmPoisson}
\usage{
spSwarmPoisson(spSwarm, edge.cutoff, min.pval = 1, min.num.edges = 0, ...)
}
\arguments{
\item{spSwarm}{An spSwarm object.}

\item{edge.cutoff}{The minimum fraction to consider (?).}

\item{min.pval}{Minimum p-value to report.}

\item{min.num.edges}{Minimum number of observed edges to report a connection.}

\item{...}{additional arguments to pass on}
}
\value{
spSwarm connection strengths and p-values.
}
\description{
Subtitle
}
\details{
Description
}
\examples{

#use demo data
output <- spSwarmPoisson(testSwa, 1/10.5)


}
\author{
Jason T. Serviss
}
\keyword{spSwarmPoisson}
