% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spUnsupervised.R
\name{spTopMax}
\alias{spTopMax}
\alias{spTopMax}
\title{spTopMax}
\usage{
spTopMax(spCounts, n)
}
\arguments{
\item{spCounts}{An spCounts object.}

\item{n}{Number of genes to select. If n > dim(data)[1] all data is returned.}
}
\value{
A numeric vector containing the indices of selected genes.
}
\description{
Facilitates gene selection prior to unsupervised clustering.
}
\details{
Returns the index for the n genes (rows) with the maximum
expression in the spCounts object. The expression matrix in
the counts.cpm slot is used for the calculation.
}
\examples{

s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])
selected <- spTopMax(cObjSng, 10)

}
\author{
Jason T. Serviss
}
\keyword{spTopMax}
