% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-classes.R, R/initialize-methods.R, R/show-methods.R, R/unsupervised.R
\docType{class}
\name{spUnsupervised-class}
\alias{.Counts}
\alias{getData,spUnsupervised-method}
\alias{initialize,spUnsupervised-method}
\alias{show,spUnsupervised-method}
\alias{spUnsupervised}
\alias{spUnsupervised,spCounts-method}
\alias{spUnsupervised-class}
\title{Run Tsne}
\usage{
\S4method{getData}{spUnsupervised}(x, n = NULL)

\S4method{initialize}{spUnsupervised}(.Object, ..., counts.log, dist, tsne,
  groupMeans, mclust)

\S4method{show}{spUnsupervised}(object)

spUnsupervised(spCounts, ...)

\S4method{spUnsupervised}{spCounts}(spCounts, max = 3000,
  plot.callback = NULL, k = 2, max_iter = 20000, perplexity = 10,
  initial_dims = 50, Gmax = 50, seed = 11, type = "max", ...)
}
\arguments{
\item{...}{Additional arguments to pass on}

\item{counts}{Counts object.}
}
\value{
Ercc fraction plot.
}
\description{
Subtitle
}
\details{
Imports count and count.ercc data to a sp.scRNAseq object.
}
\examples{

#use demo data

#run function

}
\author{
Jason T. Serviss
}
\keyword{spUnsupervised}

