% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-classes.R, R/initialize-methods.R, R/show-methods.R, R/spUnsupervised.R
\docType{class}
\name{spSwarm-class}
\alias{.Counts}
\alias{getData,spUnsupervised-method}
\alias{initialize,spUnsupervised-method}
\alias{show,spUnsupervised-method}
\alias{spSwarm-class}
\alias{spUnsupervised}
\alias{spUnsupervised,spCounts-method}
\title{spUnsupervised}
\usage{
\S4method{getData}{spUnsupervised}(x, n = NULL)

\S4method{initialize}{spUnsupervised}(.Object, ..., counts, counts.log,
  sampleType, tsne, tsneMeans, groupMeans, classification, selectInd)

\S4method{show}{spUnsupervised}(object)

spUnsupervised(spCounts, ...)

\S4method{spUnsupervised}{spCounts}(spCounts, theta = 0, k = 2,
  max_iter = 20000, perplexity = 10, initial_dims = 50, Gmax = 50,
  seed = 11, type = "max", max = 2000, genes = NULL, ...)
}
\arguments{
\item{...}{Additional arguments to pass on}

\item{spCounts}{spCounts object.}

\item{theta}{Passed to Rtsne.}

\item{k}{The dimensions of the resulting tsne. Passed to tsne function.}

\item{max_iter}{The max number of tSNE iterations. Passed to tsne function.}

\item{perplexity}{The perplexity argument to tsne. Passed to tsne function.}

\item{initial_dims}{The initial dimensions argument. Passed to tsne function.}

\item{Gmax}{A numeric vector of 1:Gmax passed as the "G" argument to Mclust.}

\item{seed}{Sets the seed before running tSNE.}

\item{type}{Decides if genes included are picked by their maximum expression
or maximum variance. Can be either "max", "var", or "manual". If "manual" the
genes argument must also be specified.}

\item{max}{The max number of genes to include based either on maximum expression
or maximum variance which is decided by the "type" paramater.}

\item{genes}{If type = manual, genes to be included are specified here as a character
vector. These must match the rownames in the counts variable.}
}
\value{
Ercc fraction plot.
}
\description{
Subtitle
}
\details{
Description.
}
\examples{

#use demo data

#run function

}
\author{
Jason T. Serviss
}
\keyword{spUnsupervised}

