% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All-classes.R, R/initialize-methods.R,
%   R/show-methods.R, R/spUnsupervised.R
\docType{class}
\name{spUnsupervised-class}
\alias{spUnsupervised-class}
\alias{.spUnsupervised}
\alias{getData,spUnsupervised-method}
\alias{tsne}
\alias{tsne,spUnsupervised-method}
\alias{tsneMeans}
\alias{tsneMeans,spUnsupervised-method}
\alias{groupMeans}
\alias{groupMeans,spUnsupervised-method}
\alias{classification}
\alias{classification,spUnsupervised-method}
\alias{uncertainty}
\alias{uncertainty,spUnsupervised-method}
\alias{selectInd}
\alias{selectInd,spUnsupervised-method}
\alias{tsne<-}
\alias{tsne<-,spUnsupervised-method}
\alias{tsneMeans<-}
\alias{tsneMeans<-,spUnsupervised-method}
\alias{groupMeans<-}
\alias{groupMeans<-,spUnsupervised-method}
\alias{classification<-}
\alias{classification<-,spUnsupervised-method}
\alias{uncertainty<-}
\alias{uncertainty<-,spUnsupervised-method}
\alias{selectInd<-}
\alias{selectInd<-,spUnsupervised-method}
\alias{initialize,spUnsupervised-method}
\alias{show,spUnsupervised-method}
\alias{spUnsupervised}
\alias{spUnsupervised}
\alias{spUnsupervised,spCounts-method}
\title{spUnsupervised}
\usage{
\S4method{getData}{spUnsupervised}(object, n = NULL)

tsne(object)

\S4method{tsne}{spUnsupervised}(object)

tsneMeans(object)

\S4method{tsneMeans}{spUnsupervised}(object)

groupMeans(object)

\S4method{groupMeans}{spUnsupervised}(object)

classification(object)

\S4method{classification}{spUnsupervised}(object)

uncertainty(object)

\S4method{uncertainty}{spUnsupervised}(object)

selectInd(object)

\S4method{selectInd}{spUnsupervised}(object)

tsne(object) <- value

\S4method{tsne}{spUnsupervised}(object) <- value

tsneMeans(object) <- value

\S4method{tsneMeans}{spUnsupervised}(object) <- value

groupMeans(object) <- value

\S4method{groupMeans}{spUnsupervised}(object) <- value

classification(object) <- value

\S4method{classification}{spUnsupervised}(object) <- value

uncertainty(object) <- value

\S4method{uncertainty}{spUnsupervised}(object) <- value

selectInd(object) <- value

\S4method{selectInd}{spUnsupervised}(object) <- value

\S4method{initialize}{spUnsupervised}(.Object, ..., tsne, tsneMeans, groupMeans,
  classification, uncertainty, selectInd)

\S4method{show}{spUnsupervised}(object)

spUnsupervised(spCounts, ...)

\S4method{spUnsupervised}{spCounts}(spCounts, theta = 0, k = 2,
  max_iter = 2000, perplexity = 10, initial_dims = 50, Gmax = 50,
  seed = 11, type = "max", max = 2000, genes = NULL, weighted = TRUE,
  ...)
}
\arguments{
\item{object}{spUnsupervised object.}

\item{n}{Data to extract from spUnsupervised object.}

\item{value}{Data to replace in spUnsupervised object.}

\item{.Object}{Internal object.}

\item{...}{Additional arguments to pass on.}

\item{tsne}{tSNE results.}

\item{tsneMeans}{The mean x and y positions of each cell type in the tSNE
results.}

\item{groupMeans}{The mean gene expression values for each cell type
(classificaiton).}

\item{classification}{Post-tSNE cell type classification. Typically
determined by the mclust package.}

\item{uncertainty}{Contains the uncertainty in converting a conditional
probablility from EM to a classification in model-based clustering.
Reported from mclust.}

\item{selectInd}{The indexes of the genes picked for use in spUnsupervised.}

\item{spCounts}{spCounts object with singlets only.}

\item{theta}{Passed to Rtsne.}

\item{k}{The dimensions of the resulting tsne. Passed to tsne function.}

\item{max_iter}{The max number of tSNE iterations. Passed to tsne function.}

\item{perplexity}{The perplexity argument to tsne. Passed to tsne function.}

\item{initial_dims}{The initial dimensions argument. Passed to tsne function.}

\item{Gmax}{A numeric vector of 1:Gmax passed as the "G" argument to Mclust.}

\item{seed}{Sets the seed before running tSNE.}

\item{type}{Decides if genes included are picked by their maximum expression
or maximum variance. Can be either "max", "var", "maxMean", or "manual". If
"manual" the genes argument must also be specified.}

\item{max}{The max number of genes to include based either on maximum
expression or maximum variance which is decided by the "type" paramater.}

\item{genes}{If type = manual, genes to be included are specified here as a
character vector. These must match the rownames in the counts variable.}

\item{weighted}{boolean indicating if the group means should be weighted with
uncertainty.}
}
\value{
spUnsupervised object.
}
\description{
Subtitle
}
\details{
Description.
}
\examples{

#use test data
s <- grepl("^s", colnames(testCounts))
cObjSng <- spCounts(testCounts[, s], testErcc[, s])

#run function
uObj <- spUnsupervised(cObjSng, max_iter = 100, Gmax = 6)

#run function

}
\author{
Jason T. Serviss
}
\keyword{spUnsupervised}
