% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{testCounts}
\alias{testCounts}
\title{Counts data used to demonstrate CIMseq package.}
\format{
matrix
\describe{
    \item{rownames}{Gene names}
    \item{colnames}{Samples. Singlets prefixed with "s" and multiplets "m".}
}
}
\usage{
testCounts
}
\value{
Matrix of counts.
}
\description{
Test counts data.
}
\examples{
data(testCounts)
}
