% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testErcc.R
\docType{data}
\name{testErcc}
\alias{testErcc}
\title{Simulated ERCC data used to demonstrate sp.scRNAseq package.}
\format{Matrix
\describe{
    \item{rownames}{ERCC spike-in names.}
    \item{colnames}{Samples. Singlets prefixed with "s" and multiplets "m".}
}}
\usage{
testErcc
}
\value{
Matrix of ERCC counts.
}
\description{
Test ercc.
}
\examples{
data(testData)

}
