% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyFunctions.R
\name{tidySwarm}
\alias{tidySwarm}
\title{tidySwarm}
\usage{
tidySwarm(spSwarm)
}
\arguments{
\item{spSwarm}{spSwarm; An spSwarm object.}
}
\description{
Tidy spSwarm objects.
}
\examples{

tidySwarm(test_spSwarm)

}
\author{
Jason T. Serviss
}
\keyword{tidySwarm}
