% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyFunctions.R
\name{tidyUnsupervised}
\alias{tidyUnsupervised}
\title{tidyUnsupervised}
\usage{
tidyUnsupervised(spUnsupervised)
}
\arguments{
\item{spUnsupervised;}{spUnsupervised An spUnsupervised object.}
}
\description{
Tidy spUnsupervised objects.
}
\examples{

tidyUnsupervised(testUns)

}
\author{
Jason T. Serviss
}
\keyword{tidyUnsupervised}
